update UcieTL to better match new architecture
