+++ scala/resources/vsrc/rx_clock_lane.v
@@ -1,0 +1,39 @@
+module rx_clock_lane (
+   input clkin,
+   output clkout,
+   input zen,
+   input zctl_0,
+   input zctl_1,
+   input zctl_2,
+   input zctl_3,
+   input zctl_4,
+   input zctl_5,
+   input zctl_6,
+   input zctl_7,
+   input zctl_8,
+   input zctl_9,
+   input zctl_10,
+   input zctl_11,
+   input zctl_12,
+   input zctl_13,
+   input zctl_14,
+   input zctl_15,
+   input zctl_16,
+   input zctl_17,
+   input zctl_18,
+   input zctl_19,
+   input a_en,
+   input a_pc,
+   input b_en,
+   input b_pc,
+   input sel_a,
+   input vref_sel_0,
+   input vref_sel_1,
+   input vref_sel_2,
+   input vref_sel_3,
+   input vref_sel_4,
+   input vref_sel_5,
+   input vref_sel_6
+);
+  assign clkout = clkin;
+endmodule
