+++ scala/resources/vsrc/tx_driver.v
@@ -1,0 +1,88 @@
+module tx_driver (
+   input din,
+   output dout,
+   input en,
+   input en_b,
+   input pu_ctl_0,
+   input pu_ctl_1,
+   input pu_ctl_2,
+   input pu_ctl_3,
+   input pu_ctl_4,
+   input pu_ctl_5,
+   input pu_ctl_6,
+   input pu_ctl_7,
+   input pu_ctl_8,
+   input pu_ctl_9,
+   input pu_ctl_10,
+   input pu_ctl_11,
+   input pu_ctl_12,
+   input pu_ctl_13,
+   input pu_ctl_14,
+   input pu_ctl_15,
+   input pu_ctl_16,
+   input pu_ctl_17,
+   input pu_ctl_18,
+   input pu_ctl_19,
+   input pu_ctl_20,
+   input pu_ctl_21,
+   input pu_ctl_22,
+   input pu_ctl_23,
+   input pu_ctl_24,
+   input pu_ctl_25,
+   input pu_ctl_26,
+   input pu_ctl_27,
+   input pu_ctl_28,
+   input pu_ctl_29,
+   input pu_ctl_30,
+   input pu_ctl_31,
+   input pu_ctl_32,
+   input pu_ctl_33,
+   input pu_ctl_34,
+   input pu_ctl_35,
+   input pu_ctl_36,
+   input pu_ctl_37,
+   input pu_ctl_38,
+   input pu_ctl_39,
+   input pd_ctlb_0,
+   input pd_ctlb_1,
+   input pd_ctlb_2,
+   input pd_ctlb_3,
+   input pd_ctlb_4,
+   input pd_ctlb_5,
+   input pd_ctlb_6,
+   input pd_ctlb_7,
+   input pd_ctlb_8,
+   input pd_ctlb_9,
+   input pd_ctlb_10,
+   input pd_ctlb_11,
+   input pd_ctlb_12,
+   input pd_ctlb_13,
+   input pd_ctlb_14,
+   input pd_ctlb_15,
+   input pd_ctlb_16,
+   input pd_ctlb_17,
+   input pd_ctlb_18,
+   input pd_ctlb_19,
+   input pd_ctlb_20,
+   input pd_ctlb_21,
+   input pd_ctlb_22,
+   input pd_ctlb_23,
+   input pd_ctlb_24,
+   input pd_ctlb_25,
+   input pd_ctlb_26,
+   input pd_ctlb_27,
+   input pd_ctlb_28,
+   input pd_ctlb_29,
+   input pd_ctlb_30,
+   input pd_ctlb_31,
+   input pd_ctlb_32,
+   input pd_ctlb_33,
+   input pd_ctlb_34,
+   input pd_ctlb_35,
+   input pd_ctlb_36,
+   input pd_ctlb_37,
+   input pd_ctlb_38,
+   input pd_ctlb_39
+);
+  assign dout = din;
+endmodule
