+++ scala/resources/vsrc/ucie_pll.v
@@ -1,0 +1,155 @@
+module ucie_pll #(parameter CLK_PERIOD_PS=125)(
+   input vclk_ref,
+   input vclk_refb,
+   input dref_low_0,
+   input dref_low_1,
+   input dref_low_2,
+   input dref_low_3,
+   input dref_low_4,
+   input dref_low_5,
+   input dref_low_6,
+   input dref_high_0,
+   input dref_high_1,
+   input dref_high_2,
+   input dref_high_3,
+   input dref_high_4,
+   input dref_high_5,
+   input dref_high_6,
+   input vrdac_ref,
+   input dcoarse_0,
+   input dcoarse_1,
+   input dcoarse_2,
+   input dcoarse_3,
+   input dcoarse_4,
+   input dcoarse_5,
+   input dcoarse_6,
+   input dcoarse_7,
+   input dvco_reset,
+   input dvco_resetn,
+   output vp_out,
+   output vn_out,
+   output vsar_ref_low,
+   output vsar_ref_high,
+   output vdig_clk,
+   output dfine_0,
+   output dfine_1,
+   output dfine_2,
+   output dfine_3,
+   output dfine_4,
+   output dfine_5,
+   output dfine_6,
+   output dfine_7,
+   output d_fcw_debug_0,
+   output d_fcw_debug_1,
+   output d_fcw_debug_2,
+   output d_fcw_debug_3,
+   output d_fcw_debug_4,
+   output d_fcw_debug_5,
+   output d_fcw_debug_6,
+   output d_fcw_debug_7,
+   output d_sar_debug_0,
+   output d_sar_debug_1,
+   output d_sar_debug_2,
+   output d_sar_debug_3,
+   output d_sar_debug_4,
+   output d_sar_debug_5,
+   output d_sar_debug_6,
+   output d_sar_debug_7,
+   input d_digital_reset,
+   input d_kp_0,
+   input d_kp_1,
+   input d_kp_2,
+   input d_kp_3,
+   input d_kp_4,
+   input d_kp_5,
+   input d_kp_6,
+   input d_kp_7,
+   input d_kp_8,
+   input d_kp_9,
+   input d_kp_10,
+   input d_kp_11,
+   input d_kp_12,
+   input d_kp_13,
+   input d_kp_14,
+   input d_kp_15,
+   input d_ki_0,
+   input d_ki_1,
+   input d_ki_2,
+   input d_ki_3,
+   input d_ki_4,
+   input d_ki_5,
+   input d_ki_6,
+   input d_ki_7,
+   input d_ki_8,
+   input d_ki_9,
+   input d_ki_10,
+   input d_ki_11,
+   input d_ki_12,
+   input d_ki_13,
+   input d_ki_14,
+   input d_ki_15,
+   input d_clol,
+   input d_ol_fcw_0,
+   input d_ol_fcw_1,
+   input d_ol_fcw_2,
+   input d_ol_fcw_3,
+   input d_ol_fcw_4,
+   input d_ol_fcw_5,
+   input d_ol_fcw_6,
+   input d_ol_fcw_7,
+   input d_accumulator_reset_0,
+   input d_accumulator_reset_1,
+   input d_accumulator_reset_2,
+   input d_accumulator_reset_3,
+   input d_accumulator_reset_4,
+   input d_accumulator_reset_5,
+   input d_accumulator_reset_6,
+   input d_accumulator_reset_7,
+   input d_accumulator_reset_8,
+   input d_accumulator_reset_9,
+   input d_accumulator_reset_10,
+   input d_accumulator_reset_11,
+   input d_accumulator_reset_12,
+   input d_accumulator_reset_13,
+   input d_accumulator_reset_14,
+   input d_accumulator_reset_15,
+   input d_accumulator_reset_16,
+   input d_accumulator_reset_17,
+   input d_accumulator_reset_18,
+   input d_accumulator_reset_19,
+   input d_accumulator_reset_20,
+   input d_accumulator_reset_21,
+   input d_accumulator_reset_22,
+   input d_accumulator_reset_23,
+   input d_accumulator_reset_24,
+   input d_accumulator_reset_25,
+   input d_accumulator_reset_26,
+   input d_accumulator_reset_27,
+   input d_accumulator_reset_28,
+   input d_accumulator_reset_29,
+   input d_accumulator_reset_30,
+   input d_accumulator_reset_31
+);
+  `timescale 1ps/100fs
+  wire reset = dvco_reset || !dvco_resetn || d_digital_reset;
+  reg clk = 1'b0;
+  always #(CLK_PERIOD_PS/2) clk = ~clk;
+  assign vp_out = reset ? 1'b0 : clk;
+  assign vn_out = reset ? 1'b0 : ~clk;
+  assign d_fcw_debug_0 = 1'b0;
+  assign d_fcw_debug_1 = 1'b0;
+  assign d_fcw_debug_2 = 1'b0;
+  assign d_fcw_debug_3 = 1'b0;
+  assign d_fcw_debug_4 = 1'b0;
+  assign d_fcw_debug_5 = 1'b0;
+  assign d_fcw_debug_6 = 1'b0;
+  assign d_fcw_debug_7 = 1'b0;
+  assign d_sar_debug_0 = 1'b0;
+  assign d_sar_debug_1 = 1'b0;
+  assign d_sar_debug_2 = 1'b0;
+  assign d_sar_debug_3 = 1'b0;
+  assign d_sar_debug_4 = 1'b0;
+  assign d_sar_debug_5 = 1'b0;
+  assign d_sar_debug_6 = 1'b0;
+  assign d_sar_debug_7 = 1'b0;
+endmodule
