--- scala/src/phy/AnalogBundles.scala
@@ -1,247 +1,0 @@
-package edu.berkeley.cs.uciedigital.phy
-
-import chisel3._
-import chisel3.util._
-
-class TxDinIO extends Bundle {
-  val din_0 = Input(Bool())
-  val din_1 = Input(Bool())
-  val din_2 = Input(Bool())
-  val din_3 = Input(Bool())
-  val din_4 = Input(Bool())
-  val din_5 = Input(Bool())
-  val din_6 = Input(Bool())
-  val din_7 = Input(Bool())
-  val din_8 = Input(Bool())
-  val din_9 = Input(Bool())
-  val din_10 = Input(Bool())
-  val din_11 = Input(Bool())
-  val din_12 = Input(Bool())
-  val din_13 = Input(Bool())
-  val din_14 = Input(Bool())
-  val din_15 = Input(Bool())
-  val din_16 = Input(Bool())
-  val din_17 = Input(Bool())
-  val din_18 = Input(Bool())
-  val din_19 = Input(Bool())
-  val din_20 = Input(Bool())
-  val din_21 = Input(Bool())
-  val din_22 = Input(Bool())
-  val din_23 = Input(Bool())
-  val din_24 = Input(Bool())
-  val din_25 = Input(Bool())
-  val din_26 = Input(Bool())
-  val din_27 = Input(Bool())
-  val din_28 = Input(Bool())
-  val din_29 = Input(Bool())
-  val din_30 = Input(Bool())
-  val din_31 = Input(Bool())
-}
-
-class RxDoutIO extends Bundle {
-  val dout_0 = Output(Bool())
-  val dout_1 = Output(Bool())
-  val dout_2 = Output(Bool())
-  val dout_3 = Output(Bool())
-  val dout_4 = Output(Bool())
-  val dout_5 = Output(Bool())
-  val dout_6 = Output(Bool())
-  val dout_7 = Output(Bool())
-  val dout_8 = Output(Bool())
-  val dout_9 = Output(Bool())
-  val dout_10 = Output(Bool())
-  val dout_11 = Output(Bool())
-  val dout_12 = Output(Bool())
-  val dout_13 = Output(Bool())
-  val dout_14 = Output(Bool())
-  val dout_15 = Output(Bool())
-}
-
-class TerminationControlIO extends Bundle {
-  val zctl_0 = Input(Bool())
-  val zctl_1 = Input(Bool())
-  val zctl_2 = Input(Bool())
-  val zctl_3 = Input(Bool())
-  val zctl_4 = Input(Bool())
-  val zctl_5 = Input(Bool())
-  val zctl_6 = Input(Bool())
-  val zctl_7 = Input(Bool())
-  val zctl_8 = Input(Bool())
-  val zctl_9 = Input(Bool())
-  val zctl_10 = Input(Bool())
-  val zctl_11 = Input(Bool())
-  val zctl_12 = Input(Bool())
-  val zctl_13 = Input(Bool())
-  val zctl_14 = Input(Bool())
-  val zctl_15 = Input(Bool())
-  val zctl_16 = Input(Bool())
-  val zctl_17 = Input(Bool())
-  val zctl_18 = Input(Bool())
-  val zctl_19 = Input(Bool())
-  val zctl_20 = Input(Bool())
-  val zctl_21 = Input(Bool())
-  val zctl_22 = Input(Bool())
-  val zctl_23 = Input(Bool())
-  val zctl_24 = Input(Bool())
-  val zctl_25 = Input(Bool())
-  val zctl_26 = Input(Bool())
-  val zctl_27 = Input(Bool())
-  val zctl_28 = Input(Bool())
-  val zctl_29 = Input(Bool())
-  val zctl_30 = Input(Bool())
-  val zctl_31 = Input(Bool())
-  val zctl_32 = Input(Bool())
-  val zctl_33 = Input(Bool())
-  val zctl_34 = Input(Bool())
-  val zctl_35 = Input(Bool())
-  val zctl_36 = Input(Bool())
-  val zctl_37 = Input(Bool())
-  val zctl_38 = Input(Bool())
-  val zctl_39 = Input(Bool())
-  val zctl_40 = Input(Bool())
-  val zctl_41 = Input(Bool())
-  val zctl_42 = Input(Bool())
-  val zctl_43 = Input(Bool())
-  val zctl_44 = Input(Bool())
-  val zctl_45 = Input(Bool())
-  val zctl_46 = Input(Bool())
-  val zctl_47 = Input(Bool())
-  val zctl_48 = Input(Bool())
-  val zctl_49 = Input(Bool())
-  val zctl_50 = Input(Bool())
-  val zctl_51 = Input(Bool())
-  val zctl_52 = Input(Bool())
-  val zctl_53 = Input(Bool())
-  val zctl_54 = Input(Bool())
-  val zctl_55 = Input(Bool())
-  val zctl_56 = Input(Bool())
-  val zctl_57 = Input(Bool())
-  val zctl_58 = Input(Bool())
-  val zctl_59 = Input(Bool())
-  val zctl_60 = Input(Bool())
-  val zctl_61 = Input(Bool())
-  val zctl_62 = Input(Bool())
-  val zctl_63 = Input(Bool())
-}
-
-class DriverPuControlIO extends Bundle {
-  val pu_ctl_0 = Input(Bool())
-  val pu_ctl_1 = Input(Bool())
-  val pu_ctl_2 = Input(Bool())
-  val pu_ctl_3 = Input(Bool())
-  val pu_ctl_4 = Input(Bool())
-  val pu_ctl_5 = Input(Bool())
-  val pu_ctl_6 = Input(Bool())
-  val pu_ctl_7 = Input(Bool())
-  val pu_ctl_8 = Input(Bool())
-  val pu_ctl_9 = Input(Bool())
-  val pu_ctl_10 = Input(Bool())
-  val pu_ctl_11 = Input(Bool())
-  val pu_ctl_12 = Input(Bool())
-  val pu_ctl_13 = Input(Bool())
-  val pu_ctl_14 = Input(Bool())
-  val pu_ctl_15 = Input(Bool())
-  val pu_ctl_16 = Input(Bool())
-  val pu_ctl_17 = Input(Bool())
-  val pu_ctl_18 = Input(Bool())
-  val pu_ctl_19 = Input(Bool())
-  val pu_ctl_20 = Input(Bool())
-  val pu_ctl_21 = Input(Bool())
-  val pu_ctl_22 = Input(Bool())
-  val pu_ctl_23 = Input(Bool())
-  val pu_ctl_24 = Input(Bool())
-  val pu_ctl_25 = Input(Bool())
-  val pu_ctl_26 = Input(Bool())
-  val pu_ctl_27 = Input(Bool())
-  val pu_ctl_28 = Input(Bool())
-  val pu_ctl_29 = Input(Bool())
-  val pu_ctl_30 = Input(Bool())
-  val pu_ctl_31 = Input(Bool())
-  val pu_ctl_32 = Input(Bool())
-  val pu_ctl_33 = Input(Bool())
-  val pu_ctl_34 = Input(Bool())
-  val pu_ctl_35 = Input(Bool())
-  val pu_ctl_36 = Input(Bool())
-  val pu_ctl_37 = Input(Bool())
-  val pu_ctl_38 = Input(Bool())
-  val pu_ctl_39 = Input(Bool())
-  val pu_ctl_40 = Input(Bool())
-  val pu_ctl_41 = Input(Bool())
-  val pu_ctl_42 = Input(Bool())
-  val pu_ctl_43 = Input(Bool())
-  val pu_ctl_44 = Input(Bool())
-  val pu_ctl_45 = Input(Bool())
-  val pu_ctl_46 = Input(Bool())
-  val pu_ctl_47 = Input(Bool())
-}
-
-class DriverPdControlIO extends Bundle {
-  val pd_ctlb_0 = Input(Bool())
-  val pd_ctlb_1 = Input(Bool())
-  val pd_ctlb_2 = Input(Bool())
-  val pd_ctlb_3 = Input(Bool())
-  val pd_ctlb_4 = Input(Bool())
-  val pd_ctlb_5 = Input(Bool())
-  val pd_ctlb_6 = Input(Bool())
-  val pd_ctlb_7 = Input(Bool())
-  val pd_ctlb_8 = Input(Bool())
-  val pd_ctlb_9 = Input(Bool())
-  val pd_ctlb_10 = Input(Bool())
-  val pd_ctlb_11 = Input(Bool())
-  val pd_ctlb_12 = Input(Bool())
-  val pd_ctlb_13 = Input(Bool())
-  val pd_ctlb_14 = Input(Bool())
-  val pd_ctlb_15 = Input(Bool())
-  val pd_ctlb_16 = Input(Bool())
-  val pd_ctlb_17 = Input(Bool())
-  val pd_ctlb_18 = Input(Bool())
-  val pd_ctlb_19 = Input(Bool())
-  val pd_ctlb_20 = Input(Bool())
-  val pd_ctlb_21 = Input(Bool())
-  val pd_ctlb_22 = Input(Bool())
-  val pd_ctlb_23 = Input(Bool())
-  val pd_ctlb_24 = Input(Bool())
-  val pd_ctlb_25 = Input(Bool())
-  val pd_ctlb_26 = Input(Bool())
-  val pd_ctlb_27 = Input(Bool())
-  val pd_ctlb_28 = Input(Bool())
-  val pd_ctlb_29 = Input(Bool())
-  val pd_ctlb_30 = Input(Bool())
-  val pd_ctlb_31 = Input(Bool())
-  val pd_ctlb_32 = Input(Bool())
-  val pd_ctlb_33 = Input(Bool())
-  val pd_ctlb_34 = Input(Bool())
-  val pd_ctlb_35 = Input(Bool())
-  val pd_ctlb_36 = Input(Bool())
-  val pd_ctlb_37 = Input(Bool())
-  val pd_ctlb_38 = Input(Bool())
-  val pd_ctlb_39 = Input(Bool())
-  val pd_ctlb_40 = Input(Bool())
-  val pd_ctlb_41 = Input(Bool())
-  val pd_ctlb_42 = Input(Bool())
-  val pd_ctlb_43 = Input(Bool())
-  val pd_ctlb_44 = Input(Bool())
-  val pd_ctlb_45 = Input(Bool())
-  val pd_ctlb_46 = Input(Bool())
-  val pd_ctlb_47 = Input(Bool())
-}
-
-class PiClockingControlIO extends Bundle {
-  val band_ctrl = Input(Bits(2.W))
-  val band_ctrlb = Input(Bits(2.W))
-  val mix_en = Input(Bits(16.W))
-  val mix_enb = Input(Bits(16.W))
-  val mux_en = Input(Bits(8.W))
-  val mux_enb = Input(Bits(8.W))
-}
-
-class MiscClockingControlIO extends Bundle {
-  val clkp_pen = Input(Bits(5.W))
-  val clkp_nen = Input(Bits(5.W))
-  val clkn_pen = Input(Bits(5.W))
-  val clkn_nen = Input(Bits(5.W))
-  val delay = Input(Bits(5.W))
-  val dll_en = Input(Bool())
-  val mode68 = Input(Bool())
-  val ocl = Input(Bool())
-}
