+++ scala/src/phy/PhyTest.scala
@@ -4,6 +4,8 @@ import chisel3._
 import chisel3.util._
 import chisel3.util.random._
 
+import edu.berkeley.cs.uciedigital.phy.macros._
+
 object DataMode extends ChiselEnum {
   // Send/receive finite number of bits.
   val finite = Value(0.U(1.W))
@@ -61,6 +63,7 @@ class PhyTestRegsIO(
   // =====================
   /** The test setup being targeted. */
   val testTarget = Input(TestTarget())
+  val divResetb = Input(AsyncReset())
 
   // TX CONTROL
   // =====================
@@ -136,7 +139,7 @@ class PhyTestRegsIO(
 
   // DEBUG CIRCUITRY CONTROL
   // ===========================
-  val driverctl = Input(Vec(6, new DriverControlIO))
+  val driverctl = Input(Vec(6, new DriverCtlIO))
   val txctl = Input(new TxLaneDigitalCtlIO)
   val txDebugTestMode = Input(TxTestMode())
   val txDebugDataMode = Input(DataMode())
@@ -174,8 +177,11 @@ class PhyTestIO(
 
   // PHY INTERFACE
   // ====================
-  val phy = Flipped(new PhyToDigitalIO(numLanes))
+  val tx = new DecoupledIO(new TxIO(numLanes))
+  val rx = Flipped(new DecoupledIO(new RxIO(numLanes)))
+  val sb = Flipped(new SbIO)
   val debug = Flipped(new PhyDebugIO)
+  val divResetb = Output(AsyncReset())
 
   // BUMP INTERFACE
   // ====================
@@ -187,16 +193,18 @@ class PhyTest(
     numLanes: Int = 2,
     bitCounterWidth: Int = 64,
     sim: Boolean = false
-) extends Module {
+) extends Module
+    with RequireSyncReset {
   val io = IO(new PhyTestIO(bufferDepthPerLane, numLanes, bitCounterWidth))
 
-  // TODO: Add debug RTL
+  io.divResetb := io.regs.divResetb
+
+  // TODO: Add debug RTL and custom sideband control
   io.regs.txDebugState := TxTestState.idle
   io.regs.txDebugDllCode := 0.U
   io.regs.txDebugPacketsEnqueued := 0.U
-  io.bumps.txData := false.B
-  io.phy.sbTxClk := false.B
-  io.phy.sbTxData := false.B
+  io.bumps := DontCare
+  io.sb := DontCare
 
   // General computations
   val maxBitCount = VecInit(Seq.fill(bitCounterWidth)(true.B)).asUInt
@@ -324,387 +332,349 @@ class PhyTest(
     }
   }
 
-  // TODO: Add async FIFOs here
-  for (lane <- 0 until numLanes) {
-    io.phy.tx.data(lane) := 0.U
-  }
-  io.phy.tx.valid := 0.U
-  io.phy.tx.track := 0.U
-  // Needs to always be true to send clock and track even when data isn't valid.
-  io.phy.tx.valid := true.B
+  io.tx := DontCare
 
-  io.phy.tx.clkp := io.regs.txClkP
-  io.phy.tx.clkn := io.regs.txClkN
-
-  // TODO: Add tx loopback RTL
-
-  // Unlike `io.phy.tx.valid`, only true when data is valid.
+  // Unlike `io.tx.valid`, only true when data is valid.
   val tx_valid = Wire(Bool())
   tx_valid := false.B
 
   // TODO: uncomment
   // TX logic
-  // switch(txState) {
-  //   is(TxTestState.idle) {
-  //     when(io.regs.txDataChunkIn.valid) {
-  //       for (i <- 0 until numSrams) {
-  //         when(i.U === io.regs.txDataLaneGroup) {
-  //           inputWrPorts(i) := io.regs.txDataChunkIn.bits
-  //         }
-  //       }
-  //     }
+  switch(txState) {
+    is(TxTestState.idle) {
+      when(io.regs.txDataChunkIn.valid) {
+        for (i <- 0 until numSrams) {
+          when(i.U === io.regs.txDataLaneGroup) {
+            inputWrPorts(i) := io.regs.txDataChunkIn.bits
+          }
+        }
+      }
 
-  //     when(io.regs.txExecute) {
-  //       txState := TxTestState.run
-  //     }
-  //   }
-  //   is(TxTestState.run) {
-  //     switch(io.regs.txTestMode) {
-  //       is(TxTestMode.manual) {
-  //         // Need to load first chunk ahead of time so that we can constantly send data.
-  //         when(loadedFirstChunk) {
-  //           // Increment address when packet is enqueued.
-  //           when(
-  //             Mux(
-  //               io.regs.testTarget === TestTarget.mainband,
-  //               io.phy.tx.ready,
-  //               io.phy.tx_loopback.ready
-  //             )
-  //           ) {
-  //             inputBufferAddr := (inputBufferAddrReg + 1.U) % txManualRepeatPeriod
-  //           }.otherwise {
-  //             inputBufferAddr := inputBufferAddrReg % txManualRepeatPeriod
-  //           }
-  //           // Only send the next packet if we still need to send more bits.
-  //           switch(io.regs.txDataMode) {
-  //             is(DataMode.finite) {
-  //               tx_valid := txPacketsEnqueued < io.regs.txPacketsToSend
-  //             }
-  //             is(DataMode.infinite) {
-  //               tx_valid := true.B
-  //             }
-  //           }
-  //         }.otherwise {
-  //           inputBufferAddr := 0.U
-  //           loadedFirstChunk := true.B
-  //         }
-  //       }
-  //       is(TxTestMode.lfsr) {
-  //         switch(io.regs.txDataMode) {
-  //           is(DataMode.finite) {
-  //             tx_valid := txPacketsEnqueued < io.regs.txPacketsToSend
-  //           }
-  //           is(DataMode.infinite) {
-  //             tx_valid := true.B
-  //           }
-  //         }
-  //       }
-  //     }
-  //     when(tx_valid) {
-  //       switch(io.regs.txTestMode) {
-  //         is(TxTestMode.manual) {
-  //           switch(io.regs.testTarget) {
-  //             is(TestTarget.mainband) {
-  //               for (lane <- 0 until numLanes) {
-  //                 io.phy.tx.bits.data(lane) := inputRdPorts(lane >> 2).asTypeOf(
-  //                   Vec(4, UInt(32.W))
-  //                 )(lane % 4)
-  //               }
-  //               io.phy.tx.bits.valid := inputRdPorts(numLanes >> 2).asTypeOf(
-  //                 Vec(4, UInt(32.W))
-  //               )(numLanes % 4)
-  //               io.phy.tx.bits.track := inputRdPorts((numLanes + 1) >> 2)
-  //                 .asTypeOf(Vec(4, UInt(32.W)))((numLanes + 1) % 4)
-  //             }
-  //             is(TestTarget.loopback) {
-  //               io.phy.tx_loopback.bits := inputRdPorts((numLanes + 2) >> 2)
-  //                 .asTypeOf(Vec(4, UInt(32.W)))((numLanes + 2) % 4)
-  //             }
-  //           }
-  //         }
-  //         is(TxTestMode.lfsr) {
-  //           switch(io.regs.testTarget) {
-  //             is(TestTarget.mainband) {
-  //               for (lane <- 0 until numLanes) {
-  //                 io.phy.tx.bits.data(lane) := Reverse(
-  //                   txLfsrs(lane).io.out.asUInt
-  //                 )(31, 0).asTypeOf(io.phy.tx.bits.data(lane))
-  //               }
-  //               io.phy.tx.bits.valid := io.regs.txValid
-  //               io.phy.tx.bits.track := io.regs.txTrack
-  //             }
-  //             is(TestTarget.loopback) {
-  //               io.phy.tx_loopback.bits := Reverse(
-  //                 txLfsrs(numLanes).io.out.asUInt
-  //               )(31, 0).asTypeOf(io.phy.tx_loopback.bits)
-  //             }
-  //           }
-  //         }
-  //       }
-  //     }
+      when(io.regs.txExecute) {
+        txState := TxTestState.run
+      }
+    }
+    is(TxTestState.run) {
+      switch(io.regs.txTestMode) {
+        is(TxTestMode.manual) {
+          // Need to load first chunk ahead of time so that we can constantly send data.
+          when(loadedFirstChunk) {
+            // Increment address when packet is enqueued.
+            when(
+              Mux(
+                io.regs.testTarget === TestTarget.mainband,
+                io.tx.ready,
+                // TODO: Add loopback
+                false.B // io.tx_loopback.ready
+              )
+            ) {
+              inputBufferAddr := (inputBufferAddrReg + 1.U) % txManualRepeatPeriod
+            }.otherwise {
+              inputBufferAddr := inputBufferAddrReg % txManualRepeatPeriod
+            }
+            // Only send the next packet if we still need to send more bits.
+            switch(io.regs.txDataMode) {
+              is(DataMode.finite) {
+                tx_valid := txPacketsEnqueued < io.regs.txPacketsToSend
+              }
+              is(DataMode.infinite) {
+                tx_valid := true.B
+              }
+            }
+          }.otherwise {
+            inputBufferAddr := 0.U
+            loadedFirstChunk := true.B
+          }
+        }
+        is(TxTestMode.lfsr) {
+          switch(io.regs.txDataMode) {
+            is(DataMode.finite) {
+              tx_valid := txPacketsEnqueued < io.regs.txPacketsToSend
+            }
+            is(DataMode.infinite) {
+              tx_valid := true.B
+            }
+          }
+        }
+      }
+      when(tx_valid) {
+        switch(io.regs.txTestMode) {
+          is(TxTestMode.manual) {
+            switch(io.regs.testTarget) {
+              is(TestTarget.mainband) {
+                for (lane <- 0 until numLanes) {
+                  io.tx.bits.data(lane) := inputRdPorts(lane >> 2).asTypeOf(
+                    Vec(4, UInt(32.W))
+                  )(lane % 4)
+                }
+                io.tx.bits.valid := inputRdPorts(numLanes >> 2).asTypeOf(
+                  Vec(4, UInt(32.W))
+                )(numLanes % 4)
+                io.tx.bits.track := inputRdPorts((numLanes + 1) >> 2)
+                  .asTypeOf(Vec(4, UInt(32.W)))((numLanes + 1) % 4)
+              }
+              is(TestTarget.loopback) {
+                // TODO: Add loopback
+                // io.tx_loopback.bits := inputRdPorts((numLanes + 2) >> 2)
+                //   .asTypeOf(Vec(4, UInt(32.W)))((numLanes + 2) % 4)
+              }
+            }
+          }
+          is(TxTestMode.lfsr) {
+            switch(io.regs.testTarget) {
+              is(TestTarget.mainband) {
+                for (lane <- 0 until numLanes) {
+                  io.tx.bits.data(lane) := Reverse(
+                    txLfsrs(lane).io.out.asUInt
+                  )(31, 0).asTypeOf(io.tx.bits.data(lane))
+                }
+                io.tx.bits.valid := io.regs.txValid
+                io.tx.bits.track := io.regs.txTrack
+              }
+              is(TestTarget.loopback) {
+                // TODO: Add loopback
+                // io.tx_loopback.bits := Reverse(
+                //   txLfsrs(numLanes).io.out.asUInt
+                // )(31, 0).asTypeOf(io.tx_loopback.bits)
+              }
+            }
+          }
+        }
+      }
 
-  //     when(
-  //       tx_valid && Mux(
-  //         io.regs.testTarget === TestTarget.mainband,
-  //         io.phy.tx.ready,
-  //         io.phy.tx_loopback.ready
-  //       )
-  //     ) {
-  //       txPacketsEnqueued := Mux(
-  //         txPacketsEnqueued < VecInit(
-  //           Seq.fill(txPacketsEnqueued.getWidth)(true.B)
-  //         ).asUInt,
-  //         txPacketsEnqueued + 1.U,
-  //         txPacketsEnqueued
-  //       )
-  //       inputBufferAddrReg := (inputBufferAddrReg + 1.U) % txManualRepeatPeriod
-  //       when(io.regs.txTestMode === TxTestMode.lfsr) {
-  //         for (lane <- 0 until numLanes + 1) {
-  //           txLfsrs(lane).io.increment := true.B
-  //         }
-  //       }
-  //     }
+      when(
+        tx_valid && Mux(
+          io.regs.testTarget === TestTarget.mainband,
+          io.tx.ready,
+          // TODO: Add loopback
+          false.B // io.loopback.ready
+        )
+      ) {
+        txPacketsEnqueued := Mux(
+          txPacketsEnqueued < VecInit(
+            Seq.fill(txPacketsEnqueued.getWidth)(true.B)
+          ).asUInt,
+          txPacketsEnqueued + 1.U,
+          txPacketsEnqueued
+        )
+        inputBufferAddrReg := (inputBufferAddrReg + 1.U) % txManualRepeatPeriod
+        when(io.regs.txTestMode === TxTestMode.lfsr) {
+          for (lane <- 0 until numLanes + 1) {
+            txLfsrs(lane).io.increment := true.B
+          }
+        }
+      }
 
-  //     when(
-  //       (io.regs.txTestMode === TxTestMode.lfsr || loadedFirstChunk) && !tx_valid
-  //     ) {
-  //       txState := TxTestState.done
-  //     }
-  //   }
-  //   is(TxTestState.done) {}
-  // }
-  // io.phy.tx_loopback.valid := tx_valid
+      when(
+        (io.regs.txTestMode === TxTestMode.lfsr || loadedFirstChunk) && !tx_valid
+      ) {
+        txState := TxTestState.done
+      }
+    }
+    is(TxTestState.done) {}
+  }
+  // TODO: Add loopback
+  // io.tx_loopback.valid := tx_valid
 
-  // // RX logic
+  // RX logic
 
-  // io.phy.rx.ready := true.B
-  // io.phy.rx_loopback.ready := true.B
+  io.rx.ready := true.B
+  // TODO: Add loopback
+  // io.rx_loopback.ready := true.B
 
-  // for (i <- 0 until numLanes + 2) {
-  //   val newRxBitErrors = rxBitErrors(i) +& PopCount(rxErrorMask(i))
-  //   rxBitErrors(i) := Mux(
-  //     newRxBitErrors > maxBitCount,
-  //     maxBitCount,
-  //     newRxBitErrors
-  //   )
-  // }
-  // when(io.regs.rxPauseCounters) {
-  //   rxPacketsReceivedOutput := rxPacketsReceivedOutput
-  //   rxBitErrorsOutput := rxBitErrorsOutput
-  // }.otherwise {
-  //   rxPacketsReceivedOutput := RegNext(rxPacketsReceived)
-  //   rxBitErrorsOutput := rxBitErrors
-  // }
+  for (i <- 0 until numLanes + 2) {
+    val newRxBitErrors = rxBitErrors(i) +& PopCount(rxErrorMask(i))
+    rxBitErrors(i) := Mux(
+      newRxBitErrors > maxBitCount,
+      maxBitCount,
+      newRxBitErrors
+    )
+  }
+  when(io.regs.rxPauseCounters) {
+    rxPacketsReceivedOutput := rxPacketsReceivedOutput
+    rxBitErrorsOutput := rxBitErrorsOutput
+  }.otherwise {
+    rxPacketsReceivedOutput := RegNext(rxPacketsReceived)
+    rxBitErrorsOutput := rxBitErrors
+  }
 
-  // // Dumb RX logic (starts recording as soon as valid goes high and never stops)
-  // val recordingStarted = withReset(rxReset) { RegInit(false.B) }
-  // val startRecording = Wire(Bool())
-  // val startIdx = Wire(UInt(log2Ceil(Phy.SerdesRatio).W))
-  // startRecording := false.B
-  // startIdx := 0.U
+  // Dumb RX logic (starts recording as soon as valid goes high and never stops)
+  val recordingStarted = withReset(rxReset) { RegInit(false.B) }
+  val startRecording = Wire(Bool())
+  val startIdx = Wire(UInt(log2Ceil(Phy.SerdesRatio).W))
+  startRecording := false.B
+  startIdx := 0.U
 
-  // // numLanes data lanes, 1 valid lane, 1 track lane, 1 loopback lane.
-  // val runningData = withReset(rxReset) {
-  //   RegInit(VecInit(Seq.fill(numLanes + 3)(0.U(32.W))))
-  // }
+  // numLanes data lanes, 1 valid lane, 1 track lane, 1 loopback lane.
+  val runningData = withReset(rxReset) {
+    RegInit(VecInit(Seq.fill(numLanes + 3)(0.U(32.W))))
+  }
 
-  // for (i <- 0 until numLanes + 2) {
-  //   rxErrorMask(i) := 0.U
-  // }
+  for (i <- 0 until numLanes + 2) {
+    rxErrorMask(i) := 0.U
+  }
 
-  // // Check valid streak after each packet is dequeued.
-  // when(
-  //   Mux(
-  //     io.regs.testTarget === TestTarget.mainband,
-  //     io.phy.rx.ready & io.phy.rx.valid,
-  //     io.phy.rx_loopback.ready & io.phy.rx_loopback.valid
-  //   )
-  // ) {
+  // Check valid streak after each packet is dequeued.
+  when(
+    Mux(
+      io.regs.testTarget === TestTarget.mainband,
+      io.rx.ready & io.rx.valid,
+      // TODO: Add loopback
+      false.B // io.rx_loopback.ready & io.rx_loopback.valid
+    )
+  ) {
 
-  //   // Find correct start index if recording hasn't started already.
-  //   for (i <- Phy.SerdesRatio - 1 to 0 by -1) {
-  //     val shouldStartRecording = Wire(Bool())
-  //     shouldStartRecording := false.B
-  //     switch(io.regs.testTarget) {
-  //       is(TestTarget.mainband) {
-  //         shouldStartRecording := io.phy.rx.bits.valid(i)
-  //       }
-  //       is(TestTarget.loopback) {
-  //         shouldStartRecording := io.phy.rx_loopback.bits(i)
-  //       }
-  //     }
-  //     when(!recordingStarted && shouldStartRecording) {
-  //       startRecording := true.B
-  //       startIdx := i.U
-  //       rxReceiveOffset := Phy.SerdesRatio.U - i.U
-  //     }
-  //   }
+    // Find correct start index if recording hasn't started already.
+    for (i <- Phy.SerdesRatio - 1 to 0 by -1) {
+      val shouldStartRecording = Wire(Bool())
+      shouldStartRecording := false.B
+      switch(io.regs.testTarget) {
+        is(TestTarget.mainband) {
+          shouldStartRecording := io.rx.bits.valid(i)
+        }
+        is(TestTarget.loopback) {
+          // TODO: Add loopback
+          // shouldStartRecording := io.rx_loopback.bits(i)
+        }
+      }
+      when(!recordingStarted && shouldStartRecording) {
+        startRecording := true.B
+        startIdx := i.U
+        rxReceiveOffset := Phy.SerdesRatio.U - i.U
+      }
+    }
 
-  //   recordingStarted := recordingStarted || startRecording
+    recordingStarted := recordingStarted || startRecording
 
-  //   when(!recordingStarted && !startRecording) {
-  //     // Store latest data at the beginning of the `runningData` register.
-  //     for (lane <- 0 until numLanes + 3) {
-  //       if (lane < numLanes) {
-  //         runningData(lane) := io.phy.rx.bits.data(lane)
-  //       } else if (lane == numLanes) {
-  //         runningData(lane) := io.phy.rx.bits.valid
-  //       } else if (lane == numLanes + 1) {
-  //         runningData(lane) := io.phy.rx.bits.track
-  //       } else {
-  //         runningData(lane) := io.phy.rx_loopback.bits
-  //       }
-  //     }
-  //   }.otherwise {
-  //     val fullPacketReceived =
-  //       rxReceiveOffset +& Phy.SerdesRatio.U - startIdx >= 32.U
-  //     val shouldProcessPacket =
-  //       fullPacketReceived && (io.regs.rxDataMode === DataMode.infinite || rxPacketsReceived < io.regs.rxPacketsToReceive)
-  //     shouldWrite := rxPacketsReceived < maxSramPackets && fullPacketReceived
-  //     val dataMask = Wire(UInt(64.W))
-  //     dataMask := ((1.U << (Phy.SerdesRatio.U - startIdx)) - 1.U) << rxReceiveOffset
-  //     val keepMask = Wire(UInt(64.W))
-  //     keepMask := ~dataMask
-  //     when(shouldProcessPacket) {
-  //       rxPacketsReceived := Mux(
-  //         rxPacketsReceived < VecInit(
-  //           Seq.fill(rxPacketsReceived.getWidth)(true.B)
-  //         ).asUInt,
-  //         rxPacketsReceived + 1.U,
-  //         rxPacketsReceived
-  //       )
-  //     }
-  //     for (lane <- 0 until numLanes + 3) {
-  //       val rawData = if (lane < numLanes) {
-  //         io.phy.rx.bits.data(lane)
-  //       } else if (lane == numLanes) {
-  //         io.phy.rx.bits.valid
-  //       } else if (lane == numLanes + 1) {
-  //         io.phy.rx.bits.track
-  //       } else {
-  //         io.phy.rx_loopback.bits
-  //       }
-  //       val data = Wire(UInt(64.W))
-  //       data := (rawData << rxReceiveOffset) >> startIdx
-  //       val newData = Wire(UInt(64.W))
-  //       newData := (data & dataMask) | (runningData(lane) & keepMask)
-  //       runningData(lane) := newData(31, 0)
-  //       when(fullPacketReceived) {
-  //         runningData(lane) := newData >> 32.U
-  //       }
-  //       when(shouldWrite) {
-  //         toWrite(lane >> 2)(lane % 4) := newData(31, 0)
-  //       }
+    when(!recordingStarted && !startRecording) {
+      // Store latest data at the beginning of the `runningData` register.
+      for (lane <- 0 until numLanes + 3) {
+        if (lane < numLanes) {
+          runningData(lane) := io.rx.bits.data(lane)
+        } else if (lane == numLanes) {
+          runningData(lane) := io.rx.bits.valid
+        } else if (lane == numLanes + 1) {
+          runningData(lane) := io.rx.bits.track
+        } else {
+          // TODO: Add loopback
+          // runningData(lane) := io.rx_loopback.bits
+        }
+      }
+    }.otherwise {
+      val fullPacketReceived =
+        rxReceiveOffset +& Phy.SerdesRatio.U - startIdx >= 32.U
+      val shouldProcessPacket =
+        fullPacketReceived && (io.regs.rxDataMode === DataMode.infinite || rxPacketsReceived < io.regs.rxPacketsToReceive)
+      shouldWrite := rxPacketsReceived < maxSramPackets && fullPacketReceived
+      val dataMask = Wire(UInt(64.W))
+      dataMask := ((1.U << (Phy.SerdesRatio.U - startIdx)) - 1.U) << rxReceiveOffset
+      val keepMask = Wire(UInt(64.W))
+      keepMask := ~dataMask
+      when(shouldProcessPacket) {
+        rxPacketsReceived := Mux(
+          rxPacketsReceived < VecInit(
+            Seq.fill(rxPacketsReceived.getWidth)(true.B)
+          ).asUInt,
+          rxPacketsReceived + 1.U,
+          rxPacketsReceived
+        )
+      }
+      for (lane <- 0 until numLanes + 3) {
+        val rawData = if (lane < numLanes) {
+          io.rx.bits.data(lane)
+        } else if (lane == numLanes) {
+          io.rx.bits.valid
+        } else if (lane == numLanes + 1) {
+          io.rx.bits.track
+        } else {
+          // TODO: Add loopback
+          0.U // io.rx_loopback.bits
+        }
+        val data = Wire(UInt(64.W))
+        data := (rawData << rxReceiveOffset) >> startIdx
+        val newData = Wire(UInt(64.W))
+        newData := (data & dataMask) | (runningData(lane) & keepMask)
+        runningData(lane) := newData(31, 0)
+        when(fullPacketReceived) {
+          runningData(lane) := newData >> 32.U
+        }
+        when(shouldWrite) {
+          toWrite(lane >> 2)(lane % 4) := newData(31, 0)
+        }
 
-  //       when(shouldProcessPacket) {
-  //         // Compare data against LFSR and increment LFSR.
-  //         if (lane < numLanes) {
-  //           rxLfsrs(lane).io.increment := true.B
-  //           val lfsrData = newData(31, 0)
-  //           rxErrorMask(lane) := newData(31, 0) ^ Reverse(
-  //             rxLfsrs(lane).io.out.asUInt
-  //           )(31, 0)
-  //         }
-  //         // Compare valid against intended waveform.
-  //         if (lane == numLanes) {
-  //           rxErrorMask(lane) := newData(31, 0) ^ io.regs.rxLfsrValid
-  //         }
-  //         if (lane == numLanes + 2) {
-  //           rxLfsrs(numLanes).io.increment := true.B
-  //           val lfsrData = newData(31, 0)
-  //           rxErrorMask(numLanes + 1) := newData(31, 0) ^ Reverse(
-  //             rxLfsrs(numLanes).io.out.asUInt
-  //           )(31, 0)
-  //         }
-  //       }
-  //     }
-  //   }
-  // }
+        when(shouldProcessPacket) {
+          // Compare data against LFSR and increment LFSR.
+          if (lane < numLanes) {
+            rxLfsrs(lane).io.increment := true.B
+            val lfsrData = newData(31, 0)
+            rxErrorMask(lane) := newData(31, 0) ^ Reverse(
+              rxLfsrs(lane).io.out.asUInt
+            )(31, 0)
+          }
+          // Compare valid against intended waveform.
+          if (lane == numLanes) {
+            rxErrorMask(lane) := newData(31, 0) ^ io.regs.rxLfsrValid
+          }
+          if (lane == numLanes + 2) {
+            rxLfsrs(numLanes).io.increment := true.B
+            val lfsrData = newData(31, 0)
+            rxErrorMask(numLanes + 1) := newData(31, 0) ^ Reverse(
+              rxLfsrs(numLanes).io.out.asUInt
+            )(31, 0)
+          }
+        }
+      }
+    }
+  }
 
   // TODO: Move to PHY
   // val refclkrx = Module(new ClkRx(sim))
-  // refclkrx.io.vip := io.top.phy.refClkP
-  // refclkrx.io.vin := io.top.phy.refClkN
+  // refclkrx.io.vip := io.top.refClkP
+  // refclkrx.io.vin := io.top.refClkN
   // val refclkbuf = Module(new DiffBuffer(sim))
   // refclkbuf.io.vinp := refclkrx.io.vop
   // refclkbuf.io.vinn := refclkrx.io.von
-  // io.phy.refClkP := refclkbuf.io.voutp
-  // io.phy.refClkN := refclkbuf.io.voutn
+  // io.refClkP := refclkbuf.io.voutp
+  // io.refClkN := refclkbuf.io.voutn
 
   // val bpclkrx = Module(new ClkRx(sim))
-  // bpclkrx.io.vip := io.top.phy.bypassClkP
-  // bpclkrx.io.vin := io.top.phy.bypassClkN
+  // bpclkrx.io.vip := io.top.bypassClkP
+  // bpclkrx.io.vin := io.top.bypassClkN
   // val bpclkbuf = Module(new DiffBuffer(sim))
   // bpclkbuf.io.vinp := bpclkrx.io.vop
   // bpclkbuf.io.vinn := bpclkrx.io.von
-  // io.phy.bypassClkP := bpclkbuf.io.voutp
-  // io.phy.bypassClkN := bpclkbuf.io.voutn
+  // io.bypassClkP := bpclkbuf.io.voutp
+  // io.bypassClkN := bpclkbuf.io.voutn
 
-  val testPllClkPBuf0 = Module(new SingleEndedBuffer(sim))
-  val testPllClkPBuf1 = Module(new SingleEndedBuffer(sim))
-  val testPllClkPBuf2 = Module(new SingleEndedBuffer(sim))
-  val testPllClkNBuf0 = Module(new SingleEndedBuffer(sim))
-  val testPllClkNBuf1 = Module(new SingleEndedBuffer(sim))
-  val testPllClkNBuf2 = Module(new SingleEndedBuffer(sim))
-  testPllClkPBuf0.io.vin := io.debug.testPllClkP
-  testPllClkPBuf1.io.vin := testPllClkPBuf0.io.vout
-  testPllClkPBuf2.io.vin := testPllClkPBuf1.io.vout
-  testPllClkNBuf0.io.vin := io.debug.testPllClkN
-  testPllClkNBuf1.io.vin := testPllClkNBuf0.io.vout
-  testPllClkNBuf2.io.vin := testPllClkNBuf1.io.vout
-
-  val uciePllClkBuf0 = Module(new DiffBuffer(sim))
-  val uciePllClkBuf1 = Module(new DiffBuffer(sim))
-  uciePllClkBuf0.io.vinp := io.debug.pllClkP
-  uciePllClkBuf0.io.vinn := io.debug.pllClkN
-  uciePllClkBuf1.io.vinp := uciePllClkBuf0.io.voutp
-  uciePllClkBuf1.io.vinn := uciePllClkBuf0.io.voutn
   // TODO: Hook up loopback lanes and TX data debug lane.
   // txLane.io.clkp := uciePllClkBuf1.io.voutp
   // txLane.io.clkn := uciePllClkBuf1.io.voutn
 
-  val pllClkNDiv = Module(new ClkDiv4(sim))
-  pllClkNDiv.io.clk := uciePllClkBuf0.io.voutn
-  pllClkNDiv.io.resetb := !reset.asBool
+  // val pllClkNDiv = Module(new ClkDiv4)
+  // pllClkNDiv.io.clk := uciePllClkBuf0.io.voutn
+  // pllClkNDiv.io.resetb := !reset.asBool
 
-  val rxClkPBuf0 = Module(new SingleEndedBuffer(sim))
-  val rxClkPBuf1 = Module(new SingleEndedBuffer(sim))
-  val rxClkPBuf2 = Module(new SingleEndedBuffer(sim))
-  val rxClkPBuf3 = Module(new SingleEndedBuffer(sim))
-  val rxClkNBuf0 = Module(new SingleEndedBuffer(sim))
-  val rxClkNBuf1 = Module(new SingleEndedBuffer(sim))
-  val rxClkNBuf2 = Module(new SingleEndedBuffer(sim))
-  val rxClkNBuf3 = Module(new SingleEndedBuffer(sim))
-  rxClkPBuf0.io.vin := io.debug.rxClk
-  rxClkPBuf1.io.vin := rxClkPBuf0.io.vout
-  rxClkPBuf2.io.vin := rxClkPBuf1.io.vout
-  rxClkPBuf3.io.vin := rxClkPBuf2.io.vout
-  rxClkNBuf0.io.vin := io.debug.rxClkDivided
-  rxClkNBuf1.io.vin := rxClkNBuf0.io.vout
-  rxClkNBuf2.io.vin := rxClkNBuf1.io.vout
-  rxClkNBuf3.io.vin := rxClkNBuf2.io.vout
-
-  val drivers = Seq(
-    (
-      testPllClkPBuf2.io.vout,
-      io.bumps.phy.testPllClkP,
-      "testPllClkPDriver"
-    ),
-    (
-      testPllClkNBuf2.io.vout,
-      io.bumps.phy.testPllClkN,
-      "testPllClkNDriver"
-    ),
-    (uciePllClkBuf0.io.voutp, io.bumps.phy.pllClkP, "pllClkPDriver"),
-    (pllClkNDiv.io.clkout_2, io.bumps.phy.pllClkN, "pllClkNDriver"),
-    (rxClkPBuf3.io.vout, io.bumps.phy.rxClk, "rxClkDriver"),
-    (rxClkNBuf3.io.vout, io.bumps.phy.rxClkDivided, "rxClkDivDriver")
-  ).zipWithIndex
-  for (((input, output, name), i) <- drivers) {
-    val driver = Module(new TxDriver(sim)).suggestName(name)
-    driver.io.din := input
-    output := driver.io.dout
-    // TODO: set up control signals
-    driver.io.ctl := 0.U.asTypeOf(driver.io.ctl)
-  }
+  // val drivers = Seq(
+  //   (
+  //     testPllClkPBuf2.io.vout,
+  //     io.bumps.testPllClkP,
+  //     "testPllClkPDriver"
+  //   ),
+  //   (
+  //     testPllClkNBuf2.io.vout,
+  //     io.bumps.testPllClkN,
+  //     "testPllClkNDriver"
+  //   ),
+  //   (uciePllClkBuf0.io.voutp, io.bumps.pllClkP, "pllClkPDriver"),
+  //   (pllClkNDiv.io.clkout_2, io.bumps.pllClkN, "pllClkNDriver"),
+  //   (rxClkPBuf3.io.vout, io.bumps.rxClk, "rxClkDriver"),
+  //   (rxClkNBuf3.io.vout, io.bumps.rxClkDivided, "rxClkDivDriver")
+  // ).zipWithIndex
+  // for (((input, output, name), i) <- drivers) {
+  //   val driver = Module(new TxDriver(sim)).suggestName(name)
+  //   driver.io.din := input
+  //   output := driver.io.dout
+  //   // TODO: set up control signals
+  //   driver.io.ctl := 0.U.asTypeOf(driver.io.ctl)
+  // }
 }
