+++ scala/src/phy/macros/Tx.scala
@@ -1,0 +1,1001 @@
+package edu.berkeley.cs.uciedigital.phy.macros
+
+import chisel3._
+import chisel3.util._
+
+class DriverCtlIO extends Bundle {
+  val pu_ctl = UInt(6.W)
+  val pd_ctl = UInt(6.W)
+  val en = Bool()
+  val en_b = Bool()
+}
+
+// DLL + PI + DCC control IO.
+class SkewCtlIO extends Bundle {
+  val dll_en = Bool()
+  val ocl = Bool()
+  val delay = Bits(5.W)
+  val mux_en = Bits(8.W)
+  val band_ctrl = Bits(2.W)
+  val mix_en = UInt(5.W)
+  val nen_out = UInt(5.W)
+  val pen_out = UInt(5.W)
+}
+
+class TxLaneCtlIO extends Bundle {
+  val driver = new DriverCtlIO
+  val skew = new SkewCtlIO
+}
+
+class TxLaneIO extends Bundle {
+  val dll_reset = Input(Bool())
+  val dll_resetb = Input(Bool())
+  val ser_resetb = Input(AsyncReset())
+  val clkp = Input(Clock())
+  val clkn = Input(Clock())
+  val din = Input(Bits(32.W))
+  val dout = Output(Bool())
+  val divclk = Output(Clock())
+  val dll_code = Output(UInt(5.W))
+  val ctl = Input(new TxLaneCtlIO)
+}
+
+class TxLane(implicit includeDefaultModels: Boolean = false) extends RawModule {
+  val io = IO(new TxLaneIO)
+
+  val verilogBlackBox = Module(new VerilogTxLane)
+  verilogBlackBox.io.dll_reset := io.dll_reset
+  verilogBlackBox.io.dll_resetb := io.dll_resetb
+  verilogBlackBox.io.ser_resetb := io.ser_resetb
+  verilogBlackBox.io.clkp := io.clkp
+  verilogBlackBox.io.clkn := io.clkn
+
+  verilogBlackBox.io.din_0 := io.din(0)
+  verilogBlackBox.io.din_1 := io.din(1)
+  verilogBlackBox.io.din_2 := io.din(2)
+  verilogBlackBox.io.din_3 := io.din(3)
+  verilogBlackBox.io.din_4 := io.din(4)
+  verilogBlackBox.io.din_5 := io.din(5)
+  verilogBlackBox.io.din_6 := io.din(6)
+  verilogBlackBox.io.din_7 := io.din(7)
+  verilogBlackBox.io.din_8 := io.din(8)
+  verilogBlackBox.io.din_9 := io.din(9)
+  verilogBlackBox.io.din_10 := io.din(10)
+  verilogBlackBox.io.din_11 := io.din(11)
+  verilogBlackBox.io.din_12 := io.din(12)
+  verilogBlackBox.io.din_13 := io.din(13)
+  verilogBlackBox.io.din_14 := io.din(14)
+  verilogBlackBox.io.din_15 := io.din(15)
+  verilogBlackBox.io.din_16 := io.din(16)
+  verilogBlackBox.io.din_17 := io.din(17)
+  verilogBlackBox.io.din_18 := io.din(18)
+  verilogBlackBox.io.din_19 := io.din(19)
+  verilogBlackBox.io.din_20 := io.din(20)
+  verilogBlackBox.io.din_21 := io.din(21)
+  verilogBlackBox.io.din_22 := io.din(22)
+  verilogBlackBox.io.din_23 := io.din(23)
+  verilogBlackBox.io.din_24 := io.din(24)
+  verilogBlackBox.io.din_25 := io.din(25)
+  verilogBlackBox.io.din_26 := io.din(26)
+  verilogBlackBox.io.din_27 := io.din(27)
+  verilogBlackBox.io.din_28 := io.din(28)
+  verilogBlackBox.io.din_29 := io.din(29)
+  verilogBlackBox.io.din_30 := io.din(30)
+  verilogBlackBox.io.din_31 := io.din(31)
+
+  io.dout := verilogBlackBox.io.dout
+  io.divclk := verilogBlackBox.io.divclk
+
+  val puCtlTherm = Wire(UInt(64.W))
+  puCtlTherm := (1.U << io.ctl.driver.pu_ctl) - 1.U
+  val pdCtlbTherm = Wire(UInt(64.W))
+  pdCtlbTherm := ~((1.U << io.ctl.driver.pd_ctl) - 1.U)
+  verilogBlackBox.io.pu_ctl_0 := puCtlTherm(0)
+  verilogBlackBox.io.pu_ctl_1 := puCtlTherm(1)
+  verilogBlackBox.io.pu_ctl_2 := puCtlTherm(2)
+  verilogBlackBox.io.pu_ctl_3 := puCtlTherm(3)
+  verilogBlackBox.io.pu_ctl_4 := puCtlTherm(4)
+  verilogBlackBox.io.pu_ctl_5 := puCtlTherm(5)
+  verilogBlackBox.io.pu_ctl_6 := puCtlTherm(6)
+  verilogBlackBox.io.pu_ctl_7 := puCtlTherm(7)
+  verilogBlackBox.io.pu_ctl_8 := puCtlTherm(8)
+  verilogBlackBox.io.pu_ctl_9 := puCtlTherm(9)
+  verilogBlackBox.io.pu_ctl_10 := puCtlTherm(10)
+  verilogBlackBox.io.pu_ctl_11 := puCtlTherm(11)
+  verilogBlackBox.io.pu_ctl_12 := puCtlTherm(12)
+  verilogBlackBox.io.pu_ctl_13 := puCtlTherm(13)
+  verilogBlackBox.io.pu_ctl_14 := puCtlTherm(14)
+  verilogBlackBox.io.pu_ctl_15 := puCtlTherm(15)
+  verilogBlackBox.io.pu_ctl_16 := puCtlTherm(16)
+  verilogBlackBox.io.pu_ctl_17 := puCtlTherm(17)
+  verilogBlackBox.io.pu_ctl_18 := puCtlTherm(18)
+  verilogBlackBox.io.pu_ctl_19 := puCtlTherm(19)
+  verilogBlackBox.io.pu_ctl_20 := puCtlTherm(20)
+  verilogBlackBox.io.pu_ctl_21 := puCtlTherm(21)
+  verilogBlackBox.io.pu_ctl_22 := puCtlTherm(22)
+  verilogBlackBox.io.pu_ctl_23 := puCtlTherm(23)
+  verilogBlackBox.io.pu_ctl_24 := puCtlTherm(24)
+  verilogBlackBox.io.pu_ctl_25 := puCtlTherm(25)
+  verilogBlackBox.io.pu_ctl_26 := puCtlTherm(26)
+  verilogBlackBox.io.pu_ctl_27 := puCtlTherm(27)
+  verilogBlackBox.io.pu_ctl_28 := puCtlTherm(28)
+  verilogBlackBox.io.pu_ctl_29 := puCtlTherm(29)
+  verilogBlackBox.io.pu_ctl_30 := puCtlTherm(30)
+  verilogBlackBox.io.pu_ctl_31 := puCtlTherm(31)
+  verilogBlackBox.io.pu_ctl_32 := puCtlTherm(32)
+  verilogBlackBox.io.pu_ctl_33 := puCtlTherm(33)
+  verilogBlackBox.io.pu_ctl_34 := puCtlTherm(34)
+  verilogBlackBox.io.pu_ctl_35 := puCtlTherm(35)
+  verilogBlackBox.io.pu_ctl_36 := puCtlTherm(36)
+  verilogBlackBox.io.pu_ctl_37 := puCtlTherm(37)
+  verilogBlackBox.io.pu_ctl_38 := puCtlTherm(38)
+  verilogBlackBox.io.pu_ctl_39 := puCtlTherm(39)
+  verilogBlackBox.io.pd_ctlb_0 := pdCtlbTherm(0)
+  verilogBlackBox.io.pd_ctlb_1 := pdCtlbTherm(1)
+  verilogBlackBox.io.pd_ctlb_2 := pdCtlbTherm(2)
+  verilogBlackBox.io.pd_ctlb_3 := pdCtlbTherm(3)
+  verilogBlackBox.io.pd_ctlb_4 := pdCtlbTherm(4)
+  verilogBlackBox.io.pd_ctlb_5 := pdCtlbTherm(5)
+  verilogBlackBox.io.pd_ctlb_6 := pdCtlbTherm(6)
+  verilogBlackBox.io.pd_ctlb_7 := pdCtlbTherm(7)
+  verilogBlackBox.io.pd_ctlb_8 := pdCtlbTherm(8)
+  verilogBlackBox.io.pd_ctlb_9 := pdCtlbTherm(9)
+  verilogBlackBox.io.pd_ctlb_10 := pdCtlbTherm(10)
+  verilogBlackBox.io.pd_ctlb_11 := pdCtlbTherm(11)
+  verilogBlackBox.io.pd_ctlb_12 := pdCtlbTherm(12)
+  verilogBlackBox.io.pd_ctlb_13 := pdCtlbTherm(13)
+  verilogBlackBox.io.pd_ctlb_14 := pdCtlbTherm(14)
+  verilogBlackBox.io.pd_ctlb_15 := pdCtlbTherm(15)
+  verilogBlackBox.io.pd_ctlb_16 := pdCtlbTherm(16)
+  verilogBlackBox.io.pd_ctlb_17 := pdCtlbTherm(17)
+  verilogBlackBox.io.pd_ctlb_18 := pdCtlbTherm(18)
+  verilogBlackBox.io.pd_ctlb_19 := pdCtlbTherm(19)
+  verilogBlackBox.io.pd_ctlb_20 := pdCtlbTherm(20)
+  verilogBlackBox.io.pd_ctlb_21 := pdCtlbTherm(21)
+  verilogBlackBox.io.pd_ctlb_22 := pdCtlbTherm(22)
+  verilogBlackBox.io.pd_ctlb_23 := pdCtlbTherm(23)
+  verilogBlackBox.io.pd_ctlb_24 := pdCtlbTherm(24)
+  verilogBlackBox.io.pd_ctlb_25 := pdCtlbTherm(25)
+  verilogBlackBox.io.pd_ctlb_26 := pdCtlbTherm(26)
+  verilogBlackBox.io.pd_ctlb_27 := pdCtlbTherm(27)
+  verilogBlackBox.io.pd_ctlb_28 := pdCtlbTherm(28)
+  verilogBlackBox.io.pd_ctlb_29 := pdCtlbTherm(29)
+  verilogBlackBox.io.pd_ctlb_30 := pdCtlbTherm(30)
+  verilogBlackBox.io.pd_ctlb_31 := pdCtlbTherm(31)
+  verilogBlackBox.io.pd_ctlb_32 := pdCtlbTherm(32)
+  verilogBlackBox.io.pd_ctlb_33 := pdCtlbTherm(33)
+  verilogBlackBox.io.pd_ctlb_34 := pdCtlbTherm(34)
+  verilogBlackBox.io.pd_ctlb_35 := pdCtlbTherm(35)
+  verilogBlackBox.io.pd_ctlb_36 := pdCtlbTherm(36)
+  verilogBlackBox.io.pd_ctlb_37 := pdCtlbTherm(37)
+  verilogBlackBox.io.pd_ctlb_38 := pdCtlbTherm(38)
+  verilogBlackBox.io.pd_ctlb_39 := pdCtlbTherm(39)
+
+  verilogBlackBox.io.driver_en := io.ctl.driver.en
+  verilogBlackBox.io.driver_en_b := io.ctl.driver.en_b
+
+  verilogBlackBox.io.dll_en := io.ctl.skew.dll_en
+  verilogBlackBox.io.ocl := io.ctl.skew.ocl
+  verilogBlackBox.io.delay_0 := io.ctl.skew.delay(0)
+  verilogBlackBox.io.delay_1 := io.ctl.skew.delay(1)
+  verilogBlackBox.io.delay_2 := io.ctl.skew.delay(2)
+  verilogBlackBox.io.delay_3 := io.ctl.skew.delay(3)
+  verilogBlackBox.io.delay_4 := io.ctl.skew.delay(4)
+  verilogBlackBox.io.delayb_0 := !io.ctl.skew.delay(0)
+  verilogBlackBox.io.delayb_1 := !io.ctl.skew.delay(1)
+  verilogBlackBox.io.delayb_2 := !io.ctl.skew.delay(2)
+  verilogBlackBox.io.delayb_3 := !io.ctl.skew.delay(3)
+  verilogBlackBox.io.delayb_4 := !io.ctl.skew.delay(4)
+
+  verilogBlackBox.io.mux_en_0 := io.ctl.skew.mux_en(0)
+  verilogBlackBox.io.mux_en_1 := io.ctl.skew.mux_en(1)
+  verilogBlackBox.io.mux_en_2 := io.ctl.skew.mux_en(2)
+  verilogBlackBox.io.mux_en_3 := io.ctl.skew.mux_en(3)
+  verilogBlackBox.io.mux_en_4 := io.ctl.skew.mux_en(4)
+  verilogBlackBox.io.mux_en_5 := io.ctl.skew.mux_en(5)
+  verilogBlackBox.io.mux_en_6 := io.ctl.skew.mux_en(6)
+  verilogBlackBox.io.mux_en_7 := io.ctl.skew.mux_en(7)
+  verilogBlackBox.io.mux_enb_0 := !io.ctl.skew.mux_en(0)
+  verilogBlackBox.io.mux_enb_1 := !io.ctl.skew.mux_en(1)
+  verilogBlackBox.io.mux_enb_2 := !io.ctl.skew.mux_en(2)
+  verilogBlackBox.io.mux_enb_3 := !io.ctl.skew.mux_en(3)
+  verilogBlackBox.io.mux_enb_4 := !io.ctl.skew.mux_en(4)
+  verilogBlackBox.io.mux_enb_5 := !io.ctl.skew.mux_en(5)
+  verilogBlackBox.io.mux_enb_6 := !io.ctl.skew.mux_en(6)
+  verilogBlackBox.io.mux_enb_7 := !io.ctl.skew.mux_en(7)
+
+  verilogBlackBox.io.band_ctrl_0 := io.ctl.skew.band_ctrl(0)
+  verilogBlackBox.io.band_ctrl_1 := io.ctl.skew.band_ctrl(1)
+  verilogBlackBox.io.band_ctrlb_0 := !io.ctl.skew.band_ctrl(0)
+  verilogBlackBox.io.band_ctrlb_1 := !io.ctl.skew.band_ctrl(1)
+
+  val mixEnTherm = Wire(UInt(32.W))
+  mixEnTherm := (1.U << io.ctl.skew.mix_en) - 1.U
+  val mixEnbTherm = Wire(UInt(32.W))
+  mixEnbTherm := ~mixEnTherm
+  verilogBlackBox.io.mix_en_0 := mixEnTherm(0)
+  verilogBlackBox.io.mix_en_1 := mixEnTherm(1)
+  verilogBlackBox.io.mix_en_2 := mixEnTherm(2)
+  verilogBlackBox.io.mix_en_3 := mixEnTherm(3)
+  verilogBlackBox.io.mix_en_4 := mixEnTherm(4)
+  verilogBlackBox.io.mix_en_5 := mixEnTherm(5)
+  verilogBlackBox.io.mix_en_6 := mixEnTherm(6)
+  verilogBlackBox.io.mix_en_7 := mixEnTherm(7)
+  verilogBlackBox.io.mix_en_8 := mixEnTherm(8)
+  verilogBlackBox.io.mix_en_9 := mixEnTherm(9)
+  verilogBlackBox.io.mix_en_10 := mixEnTherm(10)
+  verilogBlackBox.io.mix_en_11 := mixEnTherm(11)
+  verilogBlackBox.io.mix_en_12 := mixEnTherm(12)
+  verilogBlackBox.io.mix_en_13 := mixEnTherm(13)
+  verilogBlackBox.io.mix_en_14 := mixEnTherm(14)
+  verilogBlackBox.io.mix_en_15 := mixEnTherm(15)
+
+  verilogBlackBox.io.mix_enb_0 := mixEnbTherm(0)
+  verilogBlackBox.io.mix_enb_1 := mixEnbTherm(1)
+  verilogBlackBox.io.mix_enb_2 := mixEnbTherm(2)
+  verilogBlackBox.io.mix_enb_3 := mixEnbTherm(3)
+  verilogBlackBox.io.mix_enb_4 := mixEnbTherm(4)
+  verilogBlackBox.io.mix_enb_5 := mixEnbTherm(5)
+  verilogBlackBox.io.mix_enb_6 := mixEnbTherm(6)
+  verilogBlackBox.io.mix_enb_7 := mixEnbTherm(7)
+  verilogBlackBox.io.mix_enb_8 := mixEnbTherm(8)
+  verilogBlackBox.io.mix_enb_9 := mixEnbTherm(9)
+  verilogBlackBox.io.mix_enb_10 := mixEnbTherm(10)
+  verilogBlackBox.io.mix_enb_11 := mixEnbTherm(11)
+  verilogBlackBox.io.mix_enb_12 := mixEnbTherm(12)
+  verilogBlackBox.io.mix_enb_13 := mixEnbTherm(13)
+  verilogBlackBox.io.mix_enb_14 := mixEnbTherm(14)
+  verilogBlackBox.io.mix_enb_15 := mixEnbTherm(15)
+
+  verilogBlackBox.io.nen_out_0 := io.ctl.skew.nen_out(0)
+  verilogBlackBox.io.nen_out_1 := io.ctl.skew.nen_out(1)
+  verilogBlackBox.io.nen_out_2 := io.ctl.skew.nen_out(2)
+  verilogBlackBox.io.nen_out_3 := io.ctl.skew.nen_out(3)
+  verilogBlackBox.io.nen_out_4 := io.ctl.skew.nen_out(4)
+  verilogBlackBox.io.nen_outb_0 := !io.ctl.skew.nen_out(0)
+  verilogBlackBox.io.nen_outb_1 := !io.ctl.skew.nen_out(1)
+  verilogBlackBox.io.nen_outb_2 := !io.ctl.skew.nen_out(2)
+  verilogBlackBox.io.nen_outb_3 := !io.ctl.skew.nen_out(3)
+  verilogBlackBox.io.nen_outb_4 := !io.ctl.skew.nen_out(4)
+
+  verilogBlackBox.io.pen_out_0 := io.ctl.skew.pen_out(0)
+  verilogBlackBox.io.pen_out_1 := io.ctl.skew.pen_out(1)
+  verilogBlackBox.io.pen_out_2 := io.ctl.skew.pen_out(2)
+  verilogBlackBox.io.pen_out_3 := io.ctl.skew.pen_out(3)
+  verilogBlackBox.io.pen_out_4 := io.ctl.skew.pen_out(4)
+  verilogBlackBox.io.pen_outb_0 := !io.ctl.skew.pen_out(0)
+  verilogBlackBox.io.pen_outb_1 := !io.ctl.skew.pen_out(1)
+  verilogBlackBox.io.pen_outb_2 := !io.ctl.skew.pen_out(2)
+  verilogBlackBox.io.pen_outb_3 := !io.ctl.skew.pen_out(3)
+  verilogBlackBox.io.pen_outb_4 := !io.ctl.skew.pen_out(4)
+
+  io.dll_code := Cat(
+    verilogBlackBox.io.dll_code_4,
+    verilogBlackBox.io.dll_code_3,
+    verilogBlackBox.io.dll_code_2,
+    verilogBlackBox.io.dll_code_1,
+    verilogBlackBox.io.dll_code_0
+  )
+}
+
+class VerilogTxLane(implicit includeDefaultModels: Boolean = false)
+    extends BlackBox
+    with HasBlackBoxResource {
+  val io = IO(new Bundle {
+    val dll_reset = Input(Bool())
+    val dll_resetb = Input(Bool())
+    val ser_resetb = Input(AsyncReset())
+    val clkp = Input(Clock())
+    val clkn = Input(Clock())
+    val din_0 = Input(Bool())
+    val din_1 = Input(Bool())
+    val din_2 = Input(Bool())
+    val din_3 = Input(Bool())
+    val din_4 = Input(Bool())
+    val din_5 = Input(Bool())
+    val din_6 = Input(Bool())
+    val din_7 = Input(Bool())
+    val din_8 = Input(Bool())
+    val din_9 = Input(Bool())
+    val din_10 = Input(Bool())
+    val din_11 = Input(Bool())
+    val din_12 = Input(Bool())
+    val din_13 = Input(Bool())
+    val din_14 = Input(Bool())
+    val din_15 = Input(Bool())
+    val din_16 = Input(Bool())
+    val din_17 = Input(Bool())
+    val din_18 = Input(Bool())
+    val din_19 = Input(Bool())
+    val din_20 = Input(Bool())
+    val din_21 = Input(Bool())
+    val din_22 = Input(Bool())
+    val din_23 = Input(Bool())
+    val din_24 = Input(Bool())
+    val din_25 = Input(Bool())
+    val din_26 = Input(Bool())
+    val din_27 = Input(Bool())
+    val din_28 = Input(Bool())
+    val din_29 = Input(Bool())
+    val din_30 = Input(Bool())
+    val din_31 = Input(Bool())
+    val dout = Output(Bool())
+    val divclk = Output(Clock())
+    val pu_ctl_0 = Input(Bool())
+    val pu_ctl_1 = Input(Bool())
+    val pu_ctl_2 = Input(Bool())
+    val pu_ctl_3 = Input(Bool())
+    val pu_ctl_4 = Input(Bool())
+    val pu_ctl_5 = Input(Bool())
+    val pu_ctl_6 = Input(Bool())
+    val pu_ctl_7 = Input(Bool())
+    val pu_ctl_8 = Input(Bool())
+    val pu_ctl_9 = Input(Bool())
+    val pu_ctl_10 = Input(Bool())
+    val pu_ctl_11 = Input(Bool())
+    val pu_ctl_12 = Input(Bool())
+    val pu_ctl_13 = Input(Bool())
+    val pu_ctl_14 = Input(Bool())
+    val pu_ctl_15 = Input(Bool())
+    val pu_ctl_16 = Input(Bool())
+    val pu_ctl_17 = Input(Bool())
+    val pu_ctl_18 = Input(Bool())
+    val pu_ctl_19 = Input(Bool())
+    val pu_ctl_20 = Input(Bool())
+    val pu_ctl_21 = Input(Bool())
+    val pu_ctl_22 = Input(Bool())
+    val pu_ctl_23 = Input(Bool())
+    val pu_ctl_24 = Input(Bool())
+    val pu_ctl_25 = Input(Bool())
+    val pu_ctl_26 = Input(Bool())
+    val pu_ctl_27 = Input(Bool())
+    val pu_ctl_28 = Input(Bool())
+    val pu_ctl_29 = Input(Bool())
+    val pu_ctl_30 = Input(Bool())
+    val pu_ctl_31 = Input(Bool())
+    val pu_ctl_32 = Input(Bool())
+    val pu_ctl_33 = Input(Bool())
+    val pu_ctl_34 = Input(Bool())
+    val pu_ctl_35 = Input(Bool())
+    val pu_ctl_36 = Input(Bool())
+    val pu_ctl_37 = Input(Bool())
+    val pu_ctl_38 = Input(Bool())
+    val pu_ctl_39 = Input(Bool())
+    val pd_ctlb_0 = Input(Bool())
+    val pd_ctlb_1 = Input(Bool())
+    val pd_ctlb_2 = Input(Bool())
+    val pd_ctlb_3 = Input(Bool())
+    val pd_ctlb_4 = Input(Bool())
+    val pd_ctlb_5 = Input(Bool())
+    val pd_ctlb_6 = Input(Bool())
+    val pd_ctlb_7 = Input(Bool())
+    val pd_ctlb_8 = Input(Bool())
+    val pd_ctlb_9 = Input(Bool())
+    val pd_ctlb_10 = Input(Bool())
+    val pd_ctlb_11 = Input(Bool())
+    val pd_ctlb_12 = Input(Bool())
+    val pd_ctlb_13 = Input(Bool())
+    val pd_ctlb_14 = Input(Bool())
+    val pd_ctlb_15 = Input(Bool())
+    val pd_ctlb_16 = Input(Bool())
+    val pd_ctlb_17 = Input(Bool())
+    val pd_ctlb_18 = Input(Bool())
+    val pd_ctlb_19 = Input(Bool())
+    val pd_ctlb_20 = Input(Bool())
+    val pd_ctlb_21 = Input(Bool())
+    val pd_ctlb_22 = Input(Bool())
+    val pd_ctlb_23 = Input(Bool())
+    val pd_ctlb_24 = Input(Bool())
+    val pd_ctlb_25 = Input(Bool())
+    val pd_ctlb_26 = Input(Bool())
+    val pd_ctlb_27 = Input(Bool())
+    val pd_ctlb_28 = Input(Bool())
+    val pd_ctlb_29 = Input(Bool())
+    val pd_ctlb_30 = Input(Bool())
+    val pd_ctlb_31 = Input(Bool())
+    val pd_ctlb_32 = Input(Bool())
+    val pd_ctlb_33 = Input(Bool())
+    val pd_ctlb_34 = Input(Bool())
+    val pd_ctlb_35 = Input(Bool())
+    val pd_ctlb_36 = Input(Bool())
+    val pd_ctlb_37 = Input(Bool())
+    val pd_ctlb_38 = Input(Bool())
+    val pd_ctlb_39 = Input(Bool())
+    val driver_en = Input(Bool())
+    val driver_en_b = Input(Bool())
+    val dll_en = Input(Bool())
+    val ocl = Input(Bool())
+    val delay_0 = Input(Bool())
+    val delay_1 = Input(Bool())
+    val delay_2 = Input(Bool())
+    val delay_3 = Input(Bool())
+    val delay_4 = Input(Bool())
+    val delayb_0 = Input(Bool())
+    val delayb_1 = Input(Bool())
+    val delayb_2 = Input(Bool())
+    val delayb_3 = Input(Bool())
+    val delayb_4 = Input(Bool())
+    val mux_en_0 = Input(Bool())
+    val mux_en_1 = Input(Bool())
+    val mux_en_2 = Input(Bool())
+    val mux_en_3 = Input(Bool())
+    val mux_en_4 = Input(Bool())
+    val mux_en_5 = Input(Bool())
+    val mux_en_6 = Input(Bool())
+    val mux_en_7 = Input(Bool())
+    val mux_enb_0 = Input(Bool())
+    val mux_enb_1 = Input(Bool())
+    val mux_enb_2 = Input(Bool())
+    val mux_enb_3 = Input(Bool())
+    val mux_enb_4 = Input(Bool())
+    val mux_enb_5 = Input(Bool())
+    val mux_enb_6 = Input(Bool())
+    val mux_enb_7 = Input(Bool())
+    val band_ctrl_0 = Input(Bool())
+    val band_ctrl_1 = Input(Bool())
+    val band_ctrlb_0 = Input(Bool())
+    val band_ctrlb_1 = Input(Bool())
+    val mix_en_0 = Input(Bool())
+    val mix_en_1 = Input(Bool())
+    val mix_en_2 = Input(Bool())
+    val mix_en_3 = Input(Bool())
+    val mix_en_4 = Input(Bool())
+    val mix_en_5 = Input(Bool())
+    val mix_en_6 = Input(Bool())
+    val mix_en_7 = Input(Bool())
+    val mix_en_8 = Input(Bool())
+    val mix_en_9 = Input(Bool())
+    val mix_en_10 = Input(Bool())
+    val mix_en_11 = Input(Bool())
+    val mix_en_12 = Input(Bool())
+    val mix_en_13 = Input(Bool())
+    val mix_en_14 = Input(Bool())
+    val mix_en_15 = Input(Bool())
+    val mix_enb_0 = Input(Bool())
+    val mix_enb_1 = Input(Bool())
+    val mix_enb_2 = Input(Bool())
+    val mix_enb_3 = Input(Bool())
+    val mix_enb_4 = Input(Bool())
+    val mix_enb_5 = Input(Bool())
+    val mix_enb_6 = Input(Bool())
+    val mix_enb_7 = Input(Bool())
+    val mix_enb_8 = Input(Bool())
+    val mix_enb_9 = Input(Bool())
+    val mix_enb_10 = Input(Bool())
+    val mix_enb_11 = Input(Bool())
+    val mix_enb_12 = Input(Bool())
+    val mix_enb_13 = Input(Bool())
+    val mix_enb_14 = Input(Bool())
+    val mix_enb_15 = Input(Bool())
+    val nen_out_0 = Input(Bool())
+    val nen_out_1 = Input(Bool())
+    val nen_out_2 = Input(Bool())
+    val nen_out_3 = Input(Bool())
+    val nen_out_4 = Input(Bool())
+    val nen_outb_0 = Input(Bool())
+    val nen_outb_1 = Input(Bool())
+    val nen_outb_2 = Input(Bool())
+    val nen_outb_3 = Input(Bool())
+    val nen_outb_4 = Input(Bool())
+    val pen_out_0 = Input(Bool())
+    val pen_out_1 = Input(Bool())
+    val pen_out_2 = Input(Bool())
+    val pen_out_3 = Input(Bool())
+    val pen_out_4 = Input(Bool())
+    val pen_outb_0 = Input(Bool())
+    val pen_outb_1 = Input(Bool())
+    val pen_outb_2 = Input(Bool())
+    val pen_outb_3 = Input(Bool())
+    val pen_outb_4 = Input(Bool())
+    val dll_code_0 = Output(Bool())
+    val dll_code_1 = Output(Bool())
+    val dll_code_2 = Output(Bool())
+    val dll_code_3 = Output(Bool())
+    val dll_code_4 = Output(Bool())
+  })
+
+  override val desiredName = "tx_lane"
+
+  if (includeDefaultModels) {
+    addResource("/vsrc/tx_lane.v")
+  }
+}
+
+class TxDriverIO extends Bundle {
+  val din = Input(Bool())
+  val dout = Output(Bool())
+  val ctl = Input(new DriverCtlIO)
+}
+
+class TxDriver(implicit includeDefaultModels: Boolean = false)
+    extends RawModule {
+  val io = IO(new TxDriverIO)
+
+  val verilogBlackBox = Module(new VerilogTxDriver)
+  verilogBlackBox.io.din := io.din
+  io.dout := verilogBlackBox.io.dout
+  val puCtlTherm = Wire(UInt(64.W))
+  puCtlTherm := (1.U << io.ctl.pu_ctl) - 1.U
+  val pdCtlbTherm = Wire(UInt(64.W))
+  pdCtlbTherm := ~((1.U << io.ctl.pd_ctl) - 1.U)
+  verilogBlackBox.io.pu_ctl_0 := puCtlTherm(0)
+  verilogBlackBox.io.pu_ctl_1 := puCtlTherm(1)
+  verilogBlackBox.io.pu_ctl_2 := puCtlTherm(2)
+  verilogBlackBox.io.pu_ctl_3 := puCtlTherm(3)
+  verilogBlackBox.io.pu_ctl_4 := puCtlTherm(4)
+  verilogBlackBox.io.pu_ctl_5 := puCtlTherm(5)
+  verilogBlackBox.io.pu_ctl_6 := puCtlTherm(6)
+  verilogBlackBox.io.pu_ctl_7 := puCtlTherm(7)
+  verilogBlackBox.io.pu_ctl_8 := puCtlTherm(8)
+  verilogBlackBox.io.pu_ctl_9 := puCtlTherm(9)
+  verilogBlackBox.io.pu_ctl_10 := puCtlTherm(10)
+  verilogBlackBox.io.pu_ctl_11 := puCtlTherm(11)
+  verilogBlackBox.io.pu_ctl_12 := puCtlTherm(12)
+  verilogBlackBox.io.pu_ctl_13 := puCtlTherm(13)
+  verilogBlackBox.io.pu_ctl_14 := puCtlTherm(14)
+  verilogBlackBox.io.pu_ctl_15 := puCtlTherm(15)
+  verilogBlackBox.io.pu_ctl_16 := puCtlTherm(16)
+  verilogBlackBox.io.pu_ctl_17 := puCtlTherm(17)
+  verilogBlackBox.io.pu_ctl_18 := puCtlTherm(18)
+  verilogBlackBox.io.pu_ctl_19 := puCtlTherm(19)
+  verilogBlackBox.io.pu_ctl_20 := puCtlTherm(20)
+  verilogBlackBox.io.pu_ctl_21 := puCtlTherm(21)
+  verilogBlackBox.io.pu_ctl_22 := puCtlTherm(22)
+  verilogBlackBox.io.pu_ctl_23 := puCtlTherm(23)
+  verilogBlackBox.io.pu_ctl_24 := puCtlTherm(24)
+  verilogBlackBox.io.pu_ctl_25 := puCtlTherm(25)
+  verilogBlackBox.io.pu_ctl_26 := puCtlTherm(26)
+  verilogBlackBox.io.pu_ctl_27 := puCtlTherm(27)
+  verilogBlackBox.io.pu_ctl_28 := puCtlTherm(28)
+  verilogBlackBox.io.pu_ctl_29 := puCtlTherm(29)
+  verilogBlackBox.io.pu_ctl_30 := puCtlTherm(30)
+  verilogBlackBox.io.pu_ctl_31 := puCtlTherm(31)
+  verilogBlackBox.io.pu_ctl_32 := puCtlTherm(32)
+  verilogBlackBox.io.pu_ctl_33 := puCtlTherm(33)
+  verilogBlackBox.io.pu_ctl_34 := puCtlTherm(34)
+  verilogBlackBox.io.pu_ctl_35 := puCtlTherm(35)
+  verilogBlackBox.io.pu_ctl_36 := puCtlTherm(36)
+  verilogBlackBox.io.pu_ctl_37 := puCtlTherm(37)
+  verilogBlackBox.io.pu_ctl_38 := puCtlTherm(38)
+  verilogBlackBox.io.pu_ctl_39 := puCtlTherm(39)
+  verilogBlackBox.io.pd_ctlb_0 := pdCtlbTherm(0)
+  verilogBlackBox.io.pd_ctlb_1 := pdCtlbTherm(1)
+  verilogBlackBox.io.pd_ctlb_2 := pdCtlbTherm(2)
+  verilogBlackBox.io.pd_ctlb_3 := pdCtlbTherm(3)
+  verilogBlackBox.io.pd_ctlb_4 := pdCtlbTherm(4)
+  verilogBlackBox.io.pd_ctlb_5 := pdCtlbTherm(5)
+  verilogBlackBox.io.pd_ctlb_6 := pdCtlbTherm(6)
+  verilogBlackBox.io.pd_ctlb_7 := pdCtlbTherm(7)
+  verilogBlackBox.io.pd_ctlb_8 := pdCtlbTherm(8)
+  verilogBlackBox.io.pd_ctlb_9 := pdCtlbTherm(9)
+  verilogBlackBox.io.pd_ctlb_10 := pdCtlbTherm(10)
+  verilogBlackBox.io.pd_ctlb_11 := pdCtlbTherm(11)
+  verilogBlackBox.io.pd_ctlb_12 := pdCtlbTherm(12)
+  verilogBlackBox.io.pd_ctlb_13 := pdCtlbTherm(13)
+  verilogBlackBox.io.pd_ctlb_14 := pdCtlbTherm(14)
+  verilogBlackBox.io.pd_ctlb_15 := pdCtlbTherm(15)
+  verilogBlackBox.io.pd_ctlb_16 := pdCtlbTherm(16)
+  verilogBlackBox.io.pd_ctlb_17 := pdCtlbTherm(17)
+  verilogBlackBox.io.pd_ctlb_18 := pdCtlbTherm(18)
+  verilogBlackBox.io.pd_ctlb_19 := pdCtlbTherm(19)
+  verilogBlackBox.io.pd_ctlb_20 := pdCtlbTherm(20)
+  verilogBlackBox.io.pd_ctlb_21 := pdCtlbTherm(21)
+  verilogBlackBox.io.pd_ctlb_22 := pdCtlbTherm(22)
+  verilogBlackBox.io.pd_ctlb_23 := pdCtlbTherm(23)
+  verilogBlackBox.io.pd_ctlb_24 := pdCtlbTherm(24)
+  verilogBlackBox.io.pd_ctlb_25 := pdCtlbTherm(25)
+  verilogBlackBox.io.pd_ctlb_26 := pdCtlbTherm(26)
+  verilogBlackBox.io.pd_ctlb_27 := pdCtlbTherm(27)
+  verilogBlackBox.io.pd_ctlb_28 := pdCtlbTherm(28)
+  verilogBlackBox.io.pd_ctlb_29 := pdCtlbTherm(29)
+  verilogBlackBox.io.pd_ctlb_30 := pdCtlbTherm(30)
+  verilogBlackBox.io.pd_ctlb_31 := pdCtlbTherm(31)
+  verilogBlackBox.io.pd_ctlb_32 := pdCtlbTherm(32)
+  verilogBlackBox.io.pd_ctlb_33 := pdCtlbTherm(33)
+  verilogBlackBox.io.pd_ctlb_34 := pdCtlbTherm(34)
+  verilogBlackBox.io.pd_ctlb_35 := pdCtlbTherm(35)
+  verilogBlackBox.io.pd_ctlb_36 := pdCtlbTherm(36)
+  verilogBlackBox.io.pd_ctlb_37 := pdCtlbTherm(37)
+  verilogBlackBox.io.pd_ctlb_38 := pdCtlbTherm(38)
+  verilogBlackBox.io.pd_ctlb_39 := pdCtlbTherm(39)
+  verilogBlackBox.io.en := io.ctl.en
+  verilogBlackBox.io.en_b := io.ctl.en_b
+}
+
+class VerilogTxDriver(implicit includeDefaultModels: Boolean = false)
+    extends BlackBox
+    with HasBlackBoxResource {
+  val io = IO(new Bundle {
+    val din = Input(Bool())
+    val dout = Output(Bool())
+    val pu_ctl_0 = Input(Bool())
+    val pu_ctl_1 = Input(Bool())
+    val pu_ctl_2 = Input(Bool())
+    val pu_ctl_3 = Input(Bool())
+    val pu_ctl_4 = Input(Bool())
+    val pu_ctl_5 = Input(Bool())
+    val pu_ctl_6 = Input(Bool())
+    val pu_ctl_7 = Input(Bool())
+    val pu_ctl_8 = Input(Bool())
+    val pu_ctl_9 = Input(Bool())
+    val pu_ctl_10 = Input(Bool())
+    val pu_ctl_11 = Input(Bool())
+    val pu_ctl_12 = Input(Bool())
+    val pu_ctl_13 = Input(Bool())
+    val pu_ctl_14 = Input(Bool())
+    val pu_ctl_15 = Input(Bool())
+    val pu_ctl_16 = Input(Bool())
+    val pu_ctl_17 = Input(Bool())
+    val pu_ctl_18 = Input(Bool())
+    val pu_ctl_19 = Input(Bool())
+    val pu_ctl_20 = Input(Bool())
+    val pu_ctl_21 = Input(Bool())
+    val pu_ctl_22 = Input(Bool())
+    val pu_ctl_23 = Input(Bool())
+    val pu_ctl_24 = Input(Bool())
+    val pu_ctl_25 = Input(Bool())
+    val pu_ctl_26 = Input(Bool())
+    val pu_ctl_27 = Input(Bool())
+    val pu_ctl_28 = Input(Bool())
+    val pu_ctl_29 = Input(Bool())
+    val pu_ctl_30 = Input(Bool())
+    val pu_ctl_31 = Input(Bool())
+    val pu_ctl_32 = Input(Bool())
+    val pu_ctl_33 = Input(Bool())
+    val pu_ctl_34 = Input(Bool())
+    val pu_ctl_35 = Input(Bool())
+    val pu_ctl_36 = Input(Bool())
+    val pu_ctl_37 = Input(Bool())
+    val pu_ctl_38 = Input(Bool())
+    val pu_ctl_39 = Input(Bool())
+    val pd_ctlb_0 = Input(Bool())
+    val pd_ctlb_1 = Input(Bool())
+    val pd_ctlb_2 = Input(Bool())
+    val pd_ctlb_3 = Input(Bool())
+    val pd_ctlb_4 = Input(Bool())
+    val pd_ctlb_5 = Input(Bool())
+    val pd_ctlb_6 = Input(Bool())
+    val pd_ctlb_7 = Input(Bool())
+    val pd_ctlb_8 = Input(Bool())
+    val pd_ctlb_9 = Input(Bool())
+    val pd_ctlb_10 = Input(Bool())
+    val pd_ctlb_11 = Input(Bool())
+    val pd_ctlb_12 = Input(Bool())
+    val pd_ctlb_13 = Input(Bool())
+    val pd_ctlb_14 = Input(Bool())
+    val pd_ctlb_15 = Input(Bool())
+    val pd_ctlb_16 = Input(Bool())
+    val pd_ctlb_17 = Input(Bool())
+    val pd_ctlb_18 = Input(Bool())
+    val pd_ctlb_19 = Input(Bool())
+    val pd_ctlb_20 = Input(Bool())
+    val pd_ctlb_21 = Input(Bool())
+    val pd_ctlb_22 = Input(Bool())
+    val pd_ctlb_23 = Input(Bool())
+    val pd_ctlb_24 = Input(Bool())
+    val pd_ctlb_25 = Input(Bool())
+    val pd_ctlb_26 = Input(Bool())
+    val pd_ctlb_27 = Input(Bool())
+    val pd_ctlb_28 = Input(Bool())
+    val pd_ctlb_29 = Input(Bool())
+    val pd_ctlb_30 = Input(Bool())
+    val pd_ctlb_31 = Input(Bool())
+    val pd_ctlb_32 = Input(Bool())
+    val pd_ctlb_33 = Input(Bool())
+    val pd_ctlb_34 = Input(Bool())
+    val pd_ctlb_35 = Input(Bool())
+    val pd_ctlb_36 = Input(Bool())
+    val pd_ctlb_37 = Input(Bool())
+    val pd_ctlb_38 = Input(Bool())
+    val pd_ctlb_39 = Input(Bool())
+    val en = Input(Bool())
+    val en_b = Input(Bool())
+  })
+
+  override val desiredName = "tx_driver"
+
+  if (includeDefaultModels) {
+    addResource("/vsrc/tx_driver.v")
+  }
+}
+
+class PllCtlIO extends Bundle {
+  val dref_low = UInt(7.W)
+  val dref_high = UInt(7.W)
+  val dcoarse = UInt(8.W)
+  val d_kp = UInt(16.W)
+  val d_ki = UInt(16.W)
+  val d_clol = Bool()
+  val d_ol_fcw = UInt(8.W)
+  val d_accumulator_reset = UInt(32.W)
+  val vco_reset = Bool()
+  val digital_reset = Bool()
+}
+
+class PllDebugOutIO extends Bundle {
+  val d_fcw_debug = Output(UInt(8.W))
+  val d_sar_debug = Output(UInt(8.W))
+}
+
+class PllIO extends Bundle {
+  val vclk_ref = Input(Bool())
+  val vclk_refb = Input(Bool())
+  val vrdac_ref = Input(Bool())
+  val vp_out = Output(Bool())
+  val vn_out = Output(Bool())
+  val ctl = Input(new PllCtlIO)
+  val debug = new PllDebugOutIO
+}
+
+class Pll(implicit includeDefaultModels: Boolean = false) extends RawModule {
+  val io = IO(new PllIO)
+
+  val verilogBlackBox = Module(new VerilogPll)
+  verilogBlackBox.io.vclk_ref := io.vclk_ref
+  verilogBlackBox.io.vclk_refb := io.vclk_refb
+  verilogBlackBox.io.dref_low_0 := io.ctl.dref_low(0)
+  verilogBlackBox.io.dref_low_1 := io.ctl.dref_low(1)
+  verilogBlackBox.io.dref_low_2 := io.ctl.dref_low(2)
+  verilogBlackBox.io.dref_low_3 := io.ctl.dref_low(3)
+  verilogBlackBox.io.dref_low_4 := io.ctl.dref_low(4)
+  verilogBlackBox.io.dref_low_5 := io.ctl.dref_low(5)
+  verilogBlackBox.io.dref_low_6 := io.ctl.dref_low(6)
+  verilogBlackBox.io.dref_high_0 := io.ctl.dref_high(0)
+  verilogBlackBox.io.dref_high_1 := io.ctl.dref_high(1)
+  verilogBlackBox.io.dref_high_2 := io.ctl.dref_high(2)
+  verilogBlackBox.io.dref_high_3 := io.ctl.dref_high(3)
+  verilogBlackBox.io.dref_high_4 := io.ctl.dref_high(4)
+  verilogBlackBox.io.dref_high_5 := io.ctl.dref_high(5)
+  verilogBlackBox.io.dref_high_6 := io.ctl.dref_high(6)
+  verilogBlackBox.io.vrdac_ref := io.vrdac_ref
+  verilogBlackBox.io.dcoarse_0 := io.ctl.dcoarse(0)
+  verilogBlackBox.io.dcoarse_1 := io.ctl.dcoarse(1)
+  verilogBlackBox.io.dcoarse_2 := io.ctl.dcoarse(2)
+  verilogBlackBox.io.dcoarse_3 := io.ctl.dcoarse(3)
+  verilogBlackBox.io.dcoarse_4 := io.ctl.dcoarse(4)
+  verilogBlackBox.io.dcoarse_5 := io.ctl.dcoarse(5)
+  verilogBlackBox.io.dcoarse_6 := io.ctl.dcoarse(6)
+  verilogBlackBox.io.dcoarse_7 := io.ctl.dcoarse(7)
+  verilogBlackBox.io.dvco_reset := io.ctl.vco_reset
+  verilogBlackBox.io.dvco_resetn := !io.ctl.vco_reset
+  io.vp_out := verilogBlackBox.io.vp_out
+  io.vn_out := verilogBlackBox.io.vn_out
+  io.debug.d_fcw_debug := Cat(
+    verilogBlackBox.io.d_fcw_debug_7,
+    verilogBlackBox.io.d_fcw_debug_6,
+    verilogBlackBox.io.d_fcw_debug_5,
+    verilogBlackBox.io.d_fcw_debug_4,
+    verilogBlackBox.io.d_fcw_debug_3,
+    verilogBlackBox.io.d_fcw_debug_2,
+    verilogBlackBox.io.d_fcw_debug_1,
+    verilogBlackBox.io.d_fcw_debug_0
+  )
+  io.debug.d_sar_debug := Cat(
+    verilogBlackBox.io.d_sar_debug_7,
+    verilogBlackBox.io.d_sar_debug_6,
+    verilogBlackBox.io.d_sar_debug_5,
+    verilogBlackBox.io.d_sar_debug_4,
+    verilogBlackBox.io.d_sar_debug_3,
+    verilogBlackBox.io.d_sar_debug_2,
+    verilogBlackBox.io.d_sar_debug_1,
+    verilogBlackBox.io.d_sar_debug_0
+  )
+  verilogBlackBox.io.d_digital_reset := io.ctl.digital_reset
+  verilogBlackBox.io.d_kp_0 := io.ctl.d_kp(0)
+  verilogBlackBox.io.d_kp_1 := io.ctl.d_kp(1)
+  verilogBlackBox.io.d_kp_2 := io.ctl.d_kp(2)
+  verilogBlackBox.io.d_kp_3 := io.ctl.d_kp(3)
+  verilogBlackBox.io.d_kp_4 := io.ctl.d_kp(4)
+  verilogBlackBox.io.d_kp_5 := io.ctl.d_kp(5)
+  verilogBlackBox.io.d_kp_6 := io.ctl.d_kp(6)
+  verilogBlackBox.io.d_kp_7 := io.ctl.d_kp(7)
+  verilogBlackBox.io.d_kp_8 := io.ctl.d_kp(8)
+  verilogBlackBox.io.d_kp_9 := io.ctl.d_kp(9)
+  verilogBlackBox.io.d_kp_10 := io.ctl.d_kp(10)
+  verilogBlackBox.io.d_kp_11 := io.ctl.d_kp(11)
+  verilogBlackBox.io.d_kp_12 := io.ctl.d_kp(12)
+  verilogBlackBox.io.d_kp_13 := io.ctl.d_kp(13)
+  verilogBlackBox.io.d_kp_14 := io.ctl.d_kp(14)
+  verilogBlackBox.io.d_kp_15 := io.ctl.d_kp(15)
+  verilogBlackBox.io.d_ki_0 := io.ctl.d_ki(0)
+  verilogBlackBox.io.d_ki_1 := io.ctl.d_ki(1)
+  verilogBlackBox.io.d_ki_2 := io.ctl.d_ki(2)
+  verilogBlackBox.io.d_ki_3 := io.ctl.d_ki(3)
+  verilogBlackBox.io.d_ki_4 := io.ctl.d_ki(4)
+  verilogBlackBox.io.d_ki_5 := io.ctl.d_ki(5)
+  verilogBlackBox.io.d_ki_6 := io.ctl.d_ki(6)
+  verilogBlackBox.io.d_ki_7 := io.ctl.d_ki(7)
+  verilogBlackBox.io.d_ki_8 := io.ctl.d_ki(8)
+  verilogBlackBox.io.d_ki_9 := io.ctl.d_ki(9)
+  verilogBlackBox.io.d_ki_10 := io.ctl.d_ki(10)
+  verilogBlackBox.io.d_ki_11 := io.ctl.d_ki(11)
+  verilogBlackBox.io.d_ki_12 := io.ctl.d_ki(12)
+  verilogBlackBox.io.d_ki_13 := io.ctl.d_ki(13)
+  verilogBlackBox.io.d_ki_14 := io.ctl.d_ki(14)
+  verilogBlackBox.io.d_ki_15 := io.ctl.d_ki(15)
+  verilogBlackBox.io.d_clol := io.ctl.d_clol
+  verilogBlackBox.io.d_ol_fcw_0 := io.ctl.d_ol_fcw(0)
+  verilogBlackBox.io.d_ol_fcw_1 := io.ctl.d_ol_fcw(1)
+  verilogBlackBox.io.d_ol_fcw_2 := io.ctl.d_ol_fcw(2)
+  verilogBlackBox.io.d_ol_fcw_3 := io.ctl.d_ol_fcw(3)
+  verilogBlackBox.io.d_ol_fcw_4 := io.ctl.d_ol_fcw(4)
+  verilogBlackBox.io.d_ol_fcw_5 := io.ctl.d_ol_fcw(5)
+  verilogBlackBox.io.d_ol_fcw_6 := io.ctl.d_ol_fcw(6)
+  verilogBlackBox.io.d_ol_fcw_7 := io.ctl.d_ol_fcw(7)
+  verilogBlackBox.io.d_accumulator_reset_0 := io.ctl.d_accumulator_reset(0)
+  verilogBlackBox.io.d_accumulator_reset_1 := io.ctl.d_accumulator_reset(1)
+  verilogBlackBox.io.d_accumulator_reset_2 := io.ctl.d_accumulator_reset(2)
+  verilogBlackBox.io.d_accumulator_reset_3 := io.ctl.d_accumulator_reset(3)
+  verilogBlackBox.io.d_accumulator_reset_4 := io.ctl.d_accumulator_reset(4)
+  verilogBlackBox.io.d_accumulator_reset_5 := io.ctl.d_accumulator_reset(5)
+  verilogBlackBox.io.d_accumulator_reset_6 := io.ctl.d_accumulator_reset(6)
+  verilogBlackBox.io.d_accumulator_reset_7 := io.ctl.d_accumulator_reset(7)
+  verilogBlackBox.io.d_accumulator_reset_8 := io.ctl.d_accumulator_reset(8)
+  verilogBlackBox.io.d_accumulator_reset_9 := io.ctl.d_accumulator_reset(9)
+  verilogBlackBox.io.d_accumulator_reset_10 := io.ctl.d_accumulator_reset(10)
+  verilogBlackBox.io.d_accumulator_reset_11 := io.ctl.d_accumulator_reset(11)
+  verilogBlackBox.io.d_accumulator_reset_12 := io.ctl.d_accumulator_reset(12)
+  verilogBlackBox.io.d_accumulator_reset_13 := io.ctl.d_accumulator_reset(13)
+  verilogBlackBox.io.d_accumulator_reset_14 := io.ctl.d_accumulator_reset(14)
+  verilogBlackBox.io.d_accumulator_reset_15 := io.ctl.d_accumulator_reset(15)
+  verilogBlackBox.io.d_accumulator_reset_16 := io.ctl.d_accumulator_reset(16)
+  verilogBlackBox.io.d_accumulator_reset_17 := io.ctl.d_accumulator_reset(17)
+  verilogBlackBox.io.d_accumulator_reset_18 := io.ctl.d_accumulator_reset(18)
+  verilogBlackBox.io.d_accumulator_reset_19 := io.ctl.d_accumulator_reset(19)
+  verilogBlackBox.io.d_accumulator_reset_20 := io.ctl.d_accumulator_reset(20)
+  verilogBlackBox.io.d_accumulator_reset_21 := io.ctl.d_accumulator_reset(21)
+  verilogBlackBox.io.d_accumulator_reset_22 := io.ctl.d_accumulator_reset(22)
+  verilogBlackBox.io.d_accumulator_reset_23 := io.ctl.d_accumulator_reset(23)
+  verilogBlackBox.io.d_accumulator_reset_24 := io.ctl.d_accumulator_reset(24)
+  verilogBlackBox.io.d_accumulator_reset_25 := io.ctl.d_accumulator_reset(25)
+  verilogBlackBox.io.d_accumulator_reset_26 := io.ctl.d_accumulator_reset(26)
+  verilogBlackBox.io.d_accumulator_reset_27 := io.ctl.d_accumulator_reset(27)
+  verilogBlackBox.io.d_accumulator_reset_28 := io.ctl.d_accumulator_reset(28)
+  verilogBlackBox.io.d_accumulator_reset_29 := io.ctl.d_accumulator_reset(29)
+  verilogBlackBox.io.d_accumulator_reset_30 := io.ctl.d_accumulator_reset(30)
+  verilogBlackBox.io.d_accumulator_reset_31 := io.ctl.d_accumulator_reset(31)
+}
+
+class VerilogPll(clkPeriodPs: Int = 125)(implicit
+    includeDefaultModels: Boolean = false
+) extends BlackBox(Map("CLK_PERIOD_PS" -> clkPeriodPs))
+    with HasBlackBoxResource {
+  // Pins marked as "leave floating" should be outputs.
+  val io = IO(new Bundle {
+    val vclk_ref = Input(Bool())
+    val vclk_refb = Input(Bool())
+    val dref_low_0 = Input(Bool())
+    val dref_low_1 = Input(Bool())
+    val dref_low_2 = Input(Bool())
+    val dref_low_3 = Input(Bool())
+    val dref_low_4 = Input(Bool())
+    val dref_low_5 = Input(Bool())
+    val dref_low_6 = Input(Bool())
+    val dref_high_0 = Input(Bool())
+    val dref_high_1 = Input(Bool())
+    val dref_high_2 = Input(Bool())
+    val dref_high_3 = Input(Bool())
+    val dref_high_4 = Input(Bool())
+    val dref_high_5 = Input(Bool())
+    val dref_high_6 = Input(Bool())
+    val vrdac_ref = Input(Bool())
+    val dcoarse_0 = Input(Bool())
+    val dcoarse_1 = Input(Bool())
+    val dcoarse_2 = Input(Bool())
+    val dcoarse_3 = Input(Bool())
+    val dcoarse_4 = Input(Bool())
+    val dcoarse_5 = Input(Bool())
+    val dcoarse_6 = Input(Bool())
+    val dcoarse_7 = Input(Bool())
+    val dvco_reset = Input(Bool())
+    val dvco_resetn = Input(Bool())
+    val vp_out = Output(Bool())
+    val vn_out = Output(Bool())
+    val vsar_ref_low = Output(Bool())
+    val vsar_ref_high = Output(Bool())
+    val vdig_clk = Output(Bool())
+    val dfine_0 = Output(Bool())
+    val dfine_1 = Output(Bool())
+    val dfine_2 = Output(Bool())
+    val dfine_3 = Output(Bool())
+    val dfine_4 = Output(Bool())
+    val dfine_5 = Output(Bool())
+    val dfine_6 = Output(Bool())
+    val dfine_7 = Output(Bool())
+    val d_fcw_debug_0 = Output(Bool())
+    val d_fcw_debug_1 = Output(Bool())
+    val d_fcw_debug_2 = Output(Bool())
+    val d_fcw_debug_3 = Output(Bool())
+    val d_fcw_debug_4 = Output(Bool())
+    val d_fcw_debug_5 = Output(Bool())
+    val d_fcw_debug_6 = Output(Bool())
+    val d_fcw_debug_7 = Output(Bool())
+    val d_sar_debug_0 = Output(Bool())
+    val d_sar_debug_1 = Output(Bool())
+    val d_sar_debug_2 = Output(Bool())
+    val d_sar_debug_3 = Output(Bool())
+    val d_sar_debug_4 = Output(Bool())
+    val d_sar_debug_5 = Output(Bool())
+    val d_sar_debug_6 = Output(Bool())
+    val d_sar_debug_7 = Output(Bool())
+    val d_digital_reset = Input(Bool())
+    val d_kp_0 = Input(Bool())
+    val d_kp_1 = Input(Bool())
+    val d_kp_2 = Input(Bool())
+    val d_kp_3 = Input(Bool())
+    val d_kp_4 = Input(Bool())
+    val d_kp_5 = Input(Bool())
+    val d_kp_6 = Input(Bool())
+    val d_kp_7 = Input(Bool())
+    val d_kp_8 = Input(Bool())
+    val d_kp_9 = Input(Bool())
+    val d_kp_10 = Input(Bool())
+    val d_kp_11 = Input(Bool())
+    val d_kp_12 = Input(Bool())
+    val d_kp_13 = Input(Bool())
+    val d_kp_14 = Input(Bool())
+    val d_kp_15 = Input(Bool())
+    val d_ki_0 = Input(Bool())
+    val d_ki_1 = Input(Bool())
+    val d_ki_2 = Input(Bool())
+    val d_ki_3 = Input(Bool())
+    val d_ki_4 = Input(Bool())
+    val d_ki_5 = Input(Bool())
+    val d_ki_6 = Input(Bool())
+    val d_ki_7 = Input(Bool())
+    val d_ki_8 = Input(Bool())
+    val d_ki_9 = Input(Bool())
+    val d_ki_10 = Input(Bool())
+    val d_ki_11 = Input(Bool())
+    val d_ki_12 = Input(Bool())
+    val d_ki_13 = Input(Bool())
+    val d_ki_14 = Input(Bool())
+    val d_ki_15 = Input(Bool())
+    val d_clol = Input(Bool())
+    val d_ol_fcw_0 = Input(Bool())
+    val d_ol_fcw_1 = Input(Bool())
+    val d_ol_fcw_2 = Input(Bool())
+    val d_ol_fcw_3 = Input(Bool())
+    val d_ol_fcw_4 = Input(Bool())
+    val d_ol_fcw_5 = Input(Bool())
+    val d_ol_fcw_6 = Input(Bool())
+    val d_ol_fcw_7 = Input(Bool())
+    val d_accumulator_reset_0 = Input(Bool())
+    val d_accumulator_reset_1 = Input(Bool())
+    val d_accumulator_reset_2 = Input(Bool())
+    val d_accumulator_reset_3 = Input(Bool())
+    val d_accumulator_reset_4 = Input(Bool())
+    val d_accumulator_reset_5 = Input(Bool())
+    val d_accumulator_reset_6 = Input(Bool())
+    val d_accumulator_reset_7 = Input(Bool())
+    val d_accumulator_reset_8 = Input(Bool())
+    val d_accumulator_reset_9 = Input(Bool())
+    val d_accumulator_reset_10 = Input(Bool())
+    val d_accumulator_reset_11 = Input(Bool())
+    val d_accumulator_reset_12 = Input(Bool())
+    val d_accumulator_reset_13 = Input(Bool())
+    val d_accumulator_reset_14 = Input(Bool())
+    val d_accumulator_reset_15 = Input(Bool())
+    val d_accumulator_reset_16 = Input(Bool())
+    val d_accumulator_reset_17 = Input(Bool())
+    val d_accumulator_reset_18 = Input(Bool())
+    val d_accumulator_reset_19 = Input(Bool())
+    val d_accumulator_reset_20 = Input(Bool())
+    val d_accumulator_reset_21 = Input(Bool())
+    val d_accumulator_reset_22 = Input(Bool())
+    val d_accumulator_reset_23 = Input(Bool())
+    val d_accumulator_reset_24 = Input(Bool())
+    val d_accumulator_reset_25 = Input(Bool())
+    val d_accumulator_reset_26 = Input(Bool())
+    val d_accumulator_reset_27 = Input(Bool())
+    val d_accumulator_reset_28 = Input(Bool())
+    val d_accumulator_reset_29 = Input(Bool())
+    val d_accumulator_reset_30 = Input(Bool())
+    val d_accumulator_reset_31 = Input(Bool())
+  })
+
+  override val desiredName = "ucie_pll"
+  if (includeDefaultModels) {
+    addResource("/vsrc/ucie_pll.v")
+  }
+}
